Predict Method for BKP or DKP Models

Generates posterior predictive summaries from a fitted BKP or DKP model at new input locations.

Usage

# S3 method for class 'BKP'
predict(object, Xnew, CI_level = 0.95, threshold = 0.5, ...)

# S3 method for class 'DKP'
predict(object, Xnew, CI_level = 0.95, ...)

Arguments

object

An object of class "BKP" or "DKP", typically returned by fit.BKP or fit.DKP.

Xnew

A numeric matrix (or vector) of new input locations where predictions are desired.

CI_level

Credible level for prediction intervals (default is 0.95, corresponding to 95% CI).

threshold

Classification threshold for binary prediction based on posterior mean (used only for BKP; default is 0.5).

...

Additional arguments passed to generic predict methods (currently not used; included for S3 method consistency).

Value

A list with the following components:

Xnew

The new input locations.

mean

BKP: Posterior mean of the success probability at each location. DKP: A matrix of posterior mean class probabilities (rows = inputs, columns = classes).

variance

BKP: Posterior variance of the success probability. DKP: A matrix of posterior variances for each class.

lower

BKP: Lower bound of the prediction interval (e.g., 2.5th percentile for 95% CI). DKP: A matrix of lower bounds for each class (e.g., 2.5th percentile).

upper

BKP: Upper bound of the prediction interval (e.g., 97.5th percentile for 95% CI). DKP: A matrix of upper bounds for each class (e.g., 97.5th percentile).

class

BKP: Predicted binary label (0 or 1), based on posterior mean and threshold, if m = 1. DKP: Predicted class label (i.e., the class with the highest posterior mean probability).

See also

Examples

# ============================================================== #
# ========================= BKP Examples ======================= #
# ============================================================== #

#-------------------------- 1D Example ---------------------------
set.seed(123)

# Define true success probability function
true_pi_fun <- function(x) {
  (1 + exp(-x^2) * cos(10 * (1 - exp(-x)) / (1 + exp(-x)))) / 2
}

n <- 30
Xbounds <- matrix(c(-2,2), nrow=1)
X <- tgp::lhs(n = n, rect = Xbounds)
true_pi <- true_pi_fun(X)
m <- sample(100, n, replace = TRUE)
y <- rbinom(n, size = m, prob = true_pi)

# Fit BKP model
model1 <- fit.BKP(X, y, m, Xbounds=Xbounds)

# Prediction
Xnew = matrix(seq(-2, 2, length = 10), ncol=1) #new data points
predict(model1, Xnew)
#> $Xnew
#>             [,1]
#>  [1,] -2.0000000
#>  [2,] -1.5555556
#>  [3,] -1.1111111
#>  [4,] -0.6666667
#>  [5,] -0.2222222
#>  [6,]  0.2222222
#>  [7,]  0.6666667
#>  [8,]  1.1111111
#>  [9,]  1.5555556
#> [10,]  2.0000000
#> 
#> $mean
#>  [1] 0.4876541 0.5318748 0.5175070 0.1956923 0.8602008 0.6057005 0.2259782
#>  [8] 0.5370140 0.5700729 0.5110550
#> 
#> $variance
#>  [1] 0.008132937 0.001834905 0.002690709 0.001167453 0.001886007 0.002055462
#>  [7] 0.001327923 0.002292589 0.002096726 0.006504397
#> 
#> $lower
#>  [1] 0.3130318 0.4476040 0.4157174 0.1332364 0.7650464 0.5152804 0.1587223
#>  [8] 0.4427054 0.4793517 0.3533662
#> 
#> $upper
#>  [1] 0.6638517 0.6152488 0.6185737 0.2667556 0.9337123 0.6926431 0.3011716
#>  [8] 0.6300169 0.6585024 0.6676235
#> 
#> $CI_level
#> [1] 0.95
#> 

#-------------------------- 2D Example ---------------------------
set.seed(123)

# Define 2D latent function and probability transformation
true_pi_fun <- function(X) {
  if(is.null(nrow(X))) X <- matrix(X, nrow=1)
  m <- 8.6928
  s <- 2.4269
  x1 <- 4*X[,1]- 2
  x2 <- 4*X[,2]- 2
  a <- 1 + (x1 + x2 + 1)^2 *
    (19- 14*x1 + 3*x1^2- 14*x2 + 6*x1*x2 + 3*x2^2)
  b <- 30 + (2*x1- 3*x2)^2 *
    (18- 32*x1 + 12*x1^2 + 48*x2- 36*x1*x2 + 27*x2^2)
  f <- log(a*b)
  f <- (f- m)/s
  return(pnorm(f))  # Transform to probability
}

n <- 100
Xbounds <- matrix(c(0, 0, 1, 1), nrow = 2)
X <- tgp::lhs(n = n, rect = Xbounds)
true_pi <- true_pi_fun(X)
m <- sample(100, n, replace = TRUE)
y <- rbinom(n, size = m, prob = true_pi)

# Fit BKP model
model2 <- fit.BKP(X, y, m, Xbounds=Xbounds)

# Prediction
x1 <- seq(Xbounds[1,1], Xbounds[1,2], length.out = 10)
x2 <- seq(Xbounds[2,1], Xbounds[2,2], length.out = 10)
Xnew <- expand.grid(x1 = x1, x2 = x2)
predict(model2, Xnew)
#> $Xnew
#>               x1        x2
#>   [1,] 0.0000000 0.0000000
#>   [2,] 0.1111111 0.0000000
#>   [3,] 0.2222222 0.0000000
#>   [4,] 0.3333333 0.0000000
#>   [5,] 0.4444444 0.0000000
#>   [6,] 0.5555556 0.0000000
#>   [7,] 0.6666667 0.0000000
#>   [8,] 0.7777778 0.0000000
#>   [9,] 0.8888889 0.0000000
#>  [10,] 1.0000000 0.0000000
#>  [11,] 0.0000000 0.1111111
#>  [12,] 0.1111111 0.1111111
#>  [13,] 0.2222222 0.1111111
#>  [14,] 0.3333333 0.1111111
#>  [15,] 0.4444444 0.1111111
#>  [16,] 0.5555556 0.1111111
#>  [17,] 0.6666667 0.1111111
#>  [18,] 0.7777778 0.1111111
#>  [19,] 0.8888889 0.1111111
#>  [20,] 1.0000000 0.1111111
#>  [21,] 0.0000000 0.2222222
#>  [22,] 0.1111111 0.2222222
#>  [23,] 0.2222222 0.2222222
#>  [24,] 0.3333333 0.2222222
#>  [25,] 0.4444444 0.2222222
#>  [26,] 0.5555556 0.2222222
#>  [27,] 0.6666667 0.2222222
#>  [28,] 0.7777778 0.2222222
#>  [29,] 0.8888889 0.2222222
#>  [30,] 1.0000000 0.2222222
#>  [31,] 0.0000000 0.3333333
#>  [32,] 0.1111111 0.3333333
#>  [33,] 0.2222222 0.3333333
#>  [34,] 0.3333333 0.3333333
#>  [35,] 0.4444444 0.3333333
#>  [36,] 0.5555556 0.3333333
#>  [37,] 0.6666667 0.3333333
#>  [38,] 0.7777778 0.3333333
#>  [39,] 0.8888889 0.3333333
#>  [40,] 1.0000000 0.3333333
#>  [41,] 0.0000000 0.4444444
#>  [42,] 0.1111111 0.4444444
#>  [43,] 0.2222222 0.4444444
#>  [44,] 0.3333333 0.4444444
#>  [45,] 0.4444444 0.4444444
#>  [46,] 0.5555556 0.4444444
#>  [47,] 0.6666667 0.4444444
#>  [48,] 0.7777778 0.4444444
#>  [49,] 0.8888889 0.4444444
#>  [50,] 1.0000000 0.4444444
#>  [51,] 0.0000000 0.5555556
#>  [52,] 0.1111111 0.5555556
#>  [53,] 0.2222222 0.5555556
#>  [54,] 0.3333333 0.5555556
#>  [55,] 0.4444444 0.5555556
#>  [56,] 0.5555556 0.5555556
#>  [57,] 0.6666667 0.5555556
#>  [58,] 0.7777778 0.5555556
#>  [59,] 0.8888889 0.5555556
#>  [60,] 1.0000000 0.5555556
#>  [61,] 0.0000000 0.6666667
#>  [62,] 0.1111111 0.6666667
#>  [63,] 0.2222222 0.6666667
#>  [64,] 0.3333333 0.6666667
#>  [65,] 0.4444444 0.6666667
#>  [66,] 0.5555556 0.6666667
#>  [67,] 0.6666667 0.6666667
#>  [68,] 0.7777778 0.6666667
#>  [69,] 0.8888889 0.6666667
#>  [70,] 1.0000000 0.6666667
#>  [71,] 0.0000000 0.7777778
#>  [72,] 0.1111111 0.7777778
#>  [73,] 0.2222222 0.7777778
#>  [74,] 0.3333333 0.7777778
#>  [75,] 0.4444444 0.7777778
#>  [76,] 0.5555556 0.7777778
#>  [77,] 0.6666667 0.7777778
#>  [78,] 0.7777778 0.7777778
#>  [79,] 0.8888889 0.7777778
#>  [80,] 1.0000000 0.7777778
#>  [81,] 0.0000000 0.8888889
#>  [82,] 0.1111111 0.8888889
#>  [83,] 0.2222222 0.8888889
#>  [84,] 0.3333333 0.8888889
#>  [85,] 0.4444444 0.8888889
#>  [86,] 0.5555556 0.8888889
#>  [87,] 0.6666667 0.8888889
#>  [88,] 0.7777778 0.8888889
#>  [89,] 0.8888889 0.8888889
#>  [90,] 1.0000000 0.8888889
#>  [91,] 0.0000000 1.0000000
#>  [92,] 0.1111111 1.0000000
#>  [93,] 0.2222222 1.0000000
#>  [94,] 0.3333333 1.0000000
#>  [95,] 0.4444444 1.0000000
#>  [96,] 0.5555556 1.0000000
#>  [97,] 0.6666667 1.0000000
#>  [98,] 0.7777778 1.0000000
#>  [99,] 0.8888889 1.0000000
#> [100,] 1.0000000 1.0000000
#> 
#> $mean
#>   [1] 0.40856287 0.37382621 0.34487621 0.37952452 0.60268956 0.61168109
#>   [7] 0.60379951 0.62802430 0.66109956 0.65567556 0.54269018 0.51383713
#>  [13] 0.36343484 0.31972327 0.49084528 0.55898709 0.58445956 0.71066277
#>  [19] 0.77310244 0.82870192 0.59231107 0.55598138 0.42309344 0.09848554
#>  [25] 0.05107661 0.11089895 0.44302909 0.54972802 0.72583900 0.83724452
#>  [31] 0.56100876 0.51328076 0.39241455 0.09202250 0.03508808 0.05556137
#>  [37] 0.14063395 0.34547025 0.60414005 0.67639039 0.81451903 0.65070401
#>  [43] 0.19266944 0.13040729 0.15287794 0.15362114 0.13666648 0.19515742
#>  [49] 0.37751280 0.48174011 0.78582230 0.65751628 0.48179289 0.41828042
#>  [55] 0.30896800 0.21665432 0.18458738 0.11826157 0.08893491 0.14417051
#>  [61] 0.53424806 0.53740961 0.60580463 0.59523734 0.48147365 0.39242340
#>  [67] 0.28454083 0.23722493 0.21411627 0.22921415 0.81801682 0.85790369
#>  [73] 0.87624923 0.88369044 0.76579839 0.53895838 0.47832624 0.33501619
#>  [79] 0.31878296 0.39496186 0.95721169 0.94122144 0.95500241 0.95051706
#>  [85] 0.88400216 0.78596054 0.67398673 0.49677261 0.48966037 0.48761844
#>  [91] 0.98497064 0.98677687 0.97577665 0.96850692 0.93591601 0.78657488
#>  [97] 0.75054243 0.63561883 0.52518817 0.50825508
#> 
#> $variance
#>   [1] 0.0312613164 0.0078141727 0.0072498397 0.0107701070 0.0043035023
#>   [6] 0.0017580223 0.0020239920 0.0015725983 0.0032655136 0.0279872684
#>  [11] 0.0114859431 0.0037108581 0.0029290823 0.0022103474 0.0022777992
#>  [16] 0.0014428125 0.0018722744 0.0011111199 0.0012030149 0.0039557896
#>  [21] 0.0032894879 0.0020914072 0.0036082049 0.0008046754 0.0002697549
#>  [26] 0.0010011140 0.0023594736 0.0020072793 0.0027324123 0.0024801209
#>  [31] 0.0055956583 0.0018900190 0.0021937105 0.0008418265 0.0002017021
#>  [36] 0.0003811635 0.0017403342 0.0045181198 0.0018704144 0.0014274644
#>  [41] 0.0024874582 0.0031485633 0.0017795678 0.0008953167 0.0005514565
#>  [46] 0.0005615293 0.0009203987 0.0018193637 0.0014458090 0.0020133614
#>  [51] 0.0013125430 0.0017416487 0.0015408031 0.0008763947 0.0010164912
#>  [56] 0.0012304177 0.0018908996 0.0014583202 0.0010845976 0.0053979032
#>  [61] 0.0050764161 0.0014701939 0.0011267051 0.0019022220 0.0027476759
#>  [66] 0.0019014181 0.0009963320 0.0009662262 0.0015874040 0.0092561377
#>  [71] 0.0036890649 0.0010694306 0.0008312632 0.0011936156 0.0038317482
#>  [76] 0.0020136296 0.0012474067 0.0012307648 0.0016833376 0.0065437572
#>  [81] 0.0006326648 0.0004955116 0.0003061625 0.0003937092 0.0013302893
#>  [86] 0.0038461483 0.0071673110 0.0049035457 0.0014104204 0.0018118358
#>  [91] 0.0001184952 0.0001121309 0.0003508370 0.0003884408 0.0016373836
#>  [96] 0.0087589861 0.0083134047 0.0211185500 0.0170571783 0.0243417729
#> 
#> $lower
#>   [1] 0.10168662 0.21031253 0.18917407 0.18950927 0.47110217 0.52806648
#>   [7] 0.51411777 0.54878129 0.54500896 0.30026363 0.33110258 0.39443889
#>  [13] 0.26100982 0.23133536 0.39764791 0.48397664 0.49855668 0.64326453
#>  [19] 0.70172062 0.68934970 0.47783844 0.46560962 0.30794755 0.05023486
#>  [25] 0.02390631 0.05688111 0.34906910 0.46130224 0.61786719 0.72874107
#>  [31] 0.41247779 0.42801991 0.30270612 0.04349037 0.01286197 0.02382231
#>  [37] 0.06946262 0.22038275 0.51797301 0.60027358 0.70753643 0.53710311
#>  [43] 0.11712014 0.07761195 0.10976654 0.11012726 0.08288356 0.11872302
#>  [49] 0.30453284 0.39423038 0.71077497 0.57355563 0.40517468 0.36086388
#>  [55] 0.24829198 0.15201058 0.10742422 0.05434203 0.03566210 0.03406741
#>  [61] 0.39389530 0.46193298 0.53913673 0.50847290 0.37937356 0.30875605
#>  [67] 0.22477114 0.17908541 0.14146342 0.07254318 0.68507993 0.78814799
#>  [73] 0.81453745 0.80803196 0.63443391 0.45056025 0.40938352 0.26809015
#>  [79] 0.24121778 0.24308437 0.89608553 0.89065909 0.91503714 0.90504514
#>  [85] 0.80366249 0.65285487 0.49846345 0.36025147 0.41625311 0.40449312
#>  [91] 0.95741153 0.95946340 0.92783647 0.92000509 0.83674290 0.57737973
#>  [97] 0.55267030 0.33301970 0.27028900 0.20924120
#> 
#> $upper
#>   [1] 0.76618308 0.55385238 0.52005883 0.59141115 0.72715308 0.69213985
#>   [7] 0.69012458 0.70397933 0.76816128 0.92905021 0.74642404 0.63244433
#>  [13] 0.47249647 0.41512380 0.58436181 0.63268149 0.66787567 0.77372498
#>  [19] 0.83734292 0.93257594 0.70195135 0.64454024 0.54261366 0.16058520
#>  [25] 0.08770170 0.17997299 0.53907383 0.63661289 0.82190395 0.92215811
#>  [31] 0.70416171 0.59815790 0.48590934 0.15619632 0.06774871 0.09952148
#>  [37] 0.23161486 0.48249628 0.68719611 0.74812012 0.90160674 0.75628262
#>  [43] 0.28167290 0.19431831 0.20160798 0.20280173 0.20124771 0.28510559
#>  [49] 0.45336613 0.56981249 0.85236458 0.73682475 0.55883959 0.47681631
#>  [55] 0.37310628 0.28913254 0.27692428 0.20258163 0.16316348 0.31497331
#>  [61] 0.67189635 0.61204282 0.67057193 0.67912941 0.58435525 0.47936814
#>  [67] 0.34832496 0.30070733 0.29707472 0.44217343 0.92047850 0.91568684
#>  [73] 0.92697798 0.94234744 0.87521908 0.62615014 0.54768161 0.40541495
#>  [79] 0.40171196 0.55804961 0.99187628 0.97681851 0.98272567 0.98173825
#>  [85] 0.94531976 0.89378544 0.82729509 0.63353858 0.56329004 0.57108655
#>  [91] 0.99836120 0.99900479 0.99795901 0.99482069 0.99015130 0.93654390
#>  [97] 0.90450752 0.88699985 0.77308469 0.80390897
#> 
#> $CI_level
#> [1] 0.95
#> 

# ============================================================== #
# ========================= DKP Examples ======================= #
# ============================================================== #

#-------------------------- 1D Example ---------------------------
set.seed(123)

# Define true class probability function (3-class)
true_pi_fun <- function(X) {
  p <- (1 + exp(-X^2) * cos(10 * (1 - exp(-X)) / (1 + exp(-X)))) / 2
  return(matrix(c(p/2, p/2, 1 - p), nrow = length(p)))
}

n <- 30
Xbounds <- matrix(c(-2, 2), nrow = 1)
X <- tgp::lhs(n = n, rect = Xbounds)
true_pi <- true_pi_fun(X)
m <- sample(100, n, replace = TRUE)

# Generate multinomial responses
Y <- t(sapply(1:n, function(i) rmultinom(1, size = m[i], prob = true_pi[i, ])))

# Fit DKP model
model1 <- fit.DKP(X, Y, Xbounds = Xbounds)

# Prediction
Xnew = matrix(seq(-2, 2, length = 10), ncol=1) #new data points
predict(model1, Xnew)
#> $Xnew
#>             [,1]
#>  [1,] -2.0000000
#>  [2,] -1.5555556
#>  [3,] -1.1111111
#>  [4,] -0.6666667
#>  [5,] -0.2222222
#>  [6,]  0.2222222
#>  [7,]  0.6666667
#>  [8,]  1.1111111
#>  [9,]  1.5555556
#> [10,]  2.0000000
#> 
#> $mean
#>            [,1]       [,2]      [,3]
#>  [1,] 0.1919444 0.27657759 0.5314780
#>  [2,] 0.3065415 0.29965790 0.3938006
#>  [3,] 0.3193745 0.27097374 0.4096518
#>  [4,] 0.1112803 0.06597305 0.8227467
#>  [5,] 0.3907534 0.43166186 0.1775848
#>  [6,] 0.2753590 0.32248139 0.4021596
#>  [7,] 0.1341606 0.09007242 0.7757670
#>  [8,] 0.3060871 0.23434495 0.4595679
#>  [9,] 0.2569524 0.27407645 0.4689711
#> [10,] 0.2974804 0.32737373 0.3751459
#> 
#> $variance
#>               [,1]         [,2]        [,3]
#>  [1,] 0.0045401038 0.0058567672 0.007288939
#>  [2,] 0.0014509468 0.0014324442 0.001629422
#>  [3,] 0.0021779426 0.0019792851 0.002423043
#>  [4,] 0.0007050916 0.0004393275 0.001039736
#>  [5,] 0.0034093227 0.0035133605 0.002091554
#>  [6,] 0.0016557605 0.0018130144 0.001995074
#>  [7,] 0.0008172041 0.0005765897 0.001223768
#>  [8,] 0.0018732590 0.0015824740 0.002190477
#>  [9,] 0.0015368291 0.0016014702 0.002004567
#> [10,] 0.0049852170 0.0052527283 0.005591729
#> 
#> $lower
#>             [,1]       [,2]       [,3]
#>  [1,] 0.07906969 0.14070981 0.36348464
#>  [2,] 0.23454336 0.22822864 0.31619251
#>  [3,] 0.23161605 0.18843642 0.31512798
#>  [4,] 0.06478408 0.03112775 0.75537129
#>  [5,] 0.27970993 0.31774730 0.09739606
#>  [6,] 0.19936718 0.24202548 0.31633381
#>  [7,] 0.08327339 0.04879376 0.70368081
#>  [8,] 0.22473409 0.16109480 0.36869548
#>  [9,] 0.18405193 0.19930662 0.38184954
#> [10,] 0.16943484 0.19424026 0.23523976
#> 
#> $upper
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3398352 0.4379332 0.6958718
#>  [2,] 0.3835768 0.3763034 0.4741746
#>  [3,] 0.4140591 0.3622901 0.5076413
#>  [4,] 0.1683209 0.1125517 0.8813517
#>  [5,] 0.5078130 0.5493403 0.2754764
#>  [6,] 0.3584707 0.4085653 0.4910881
#>  [7,] 0.1948475 0.1423104 0.8404550
#>  [8,] 0.3939771 0.3165454 0.5518040
#>  [9,] 0.3373225 0.3557905 0.5570471
#> [10,] 0.4442648 0.4764876 0.5266162
#> 
#> $CI_level
#> [1] 0.95
#> 

#-------------------------- 2D Example ---------------------------
set.seed(123)

# Define latent function and transform to 3-class probabilities
true_pi_fun <- function(X) {
  if (is.null(nrow(X))) X <- matrix(X, nrow = 1)
  m <- 8.6928; s <- 2.4269
  x1 <- 4 * X[,1] - 2
  x2 <- 4 * X[,2] - 2
  a <- 1 + (x1 + x2 + 1)^2 *
    (19 - 14*x1 + 3*x1^2 - 14*x2 + 6*x1*x2 + 3*x2^2)
  b <- 30 + (2*x1 - 3*x2)^2 *
    (18 - 32*x1 + 12*x1^2 + 48*x2 - 36*x1*x2 + 27*x2^2)
  f <- (log(a * b) - m) / s
  p <- pnorm(f)
  return(matrix(c(p/2, p/2, 1 - p), nrow = length(p)))
}

n <- 100
Xbounds <- matrix(c(0, 0, 1, 1), nrow = 2)
X <- tgp::lhs(n = n, rect = Xbounds)
true_pi <- true_pi_fun(X)
m <- sample(100, n, replace = TRUE)

# Generate multinomial responses
Y <- t(sapply(1:n, function(i) rmultinom(1, size = m[i], prob = true_pi[i, ])))

# Fit DKP model
model2 <- fit.DKP(X, Y, Xbounds = Xbounds)
print(model2)
#> --------------------------------------------------
#>            Dirichlet Kernel Process (DKP) Model        
#> --------------------------------------------------
#> Number of observations (n):  100
#> Input dimensionality (d):    2
#> Output dimensionality (q):   3
#> Kernel type:                 gaussian
#> Loss function used:          brier
#> Optimized kernel parameters: 0.1245, 0.0706
#> Minimum achieved loss:       0.00061
#> 
#> Prior specification:
#>   Noninformative prior: Dirichlet(1,...,1).
#> --------------------------------------------------

# Prediction
x1 <- seq(Xbounds[1,1], Xbounds[1,2], length.out = 10)
x2 <- seq(Xbounds[2,1], Xbounds[2,2], length.out = 10)
Xnew <- expand.grid(x1 = x1, x2 = x2)
predict(model2, Xnew)
#> $Xnew
#>               x1        x2
#>   [1,] 0.0000000 0.0000000
#>   [2,] 0.1111111 0.0000000
#>   [3,] 0.2222222 0.0000000
#>   [4,] 0.3333333 0.0000000
#>   [5,] 0.4444444 0.0000000
#>   [6,] 0.5555556 0.0000000
#>   [7,] 0.6666667 0.0000000
#>   [8,] 0.7777778 0.0000000
#>   [9,] 0.8888889 0.0000000
#>  [10,] 1.0000000 0.0000000
#>  [11,] 0.0000000 0.1111111
#>  [12,] 0.1111111 0.1111111
#>  [13,] 0.2222222 0.1111111
#>  [14,] 0.3333333 0.1111111
#>  [15,] 0.4444444 0.1111111
#>  [16,] 0.5555556 0.1111111
#>  [17,] 0.6666667 0.1111111
#>  [18,] 0.7777778 0.1111111
#>  [19,] 0.8888889 0.1111111
#>  [20,] 1.0000000 0.1111111
#>  [21,] 0.0000000 0.2222222
#>  [22,] 0.1111111 0.2222222
#>  [23,] 0.2222222 0.2222222
#>  [24,] 0.3333333 0.2222222
#>  [25,] 0.4444444 0.2222222
#>  [26,] 0.5555556 0.2222222
#>  [27,] 0.6666667 0.2222222
#>  [28,] 0.7777778 0.2222222
#>  [29,] 0.8888889 0.2222222
#>  [30,] 1.0000000 0.2222222
#>  [31,] 0.0000000 0.3333333
#>  [32,] 0.1111111 0.3333333
#>  [33,] 0.2222222 0.3333333
#>  [34,] 0.3333333 0.3333333
#>  [35,] 0.4444444 0.3333333
#>  [36,] 0.5555556 0.3333333
#>  [37,] 0.6666667 0.3333333
#>  [38,] 0.7777778 0.3333333
#>  [39,] 0.8888889 0.3333333
#>  [40,] 1.0000000 0.3333333
#>  [41,] 0.0000000 0.4444444
#>  [42,] 0.1111111 0.4444444
#>  [43,] 0.2222222 0.4444444
#>  [44,] 0.3333333 0.4444444
#>  [45,] 0.4444444 0.4444444
#>  [46,] 0.5555556 0.4444444
#>  [47,] 0.6666667 0.4444444
#>  [48,] 0.7777778 0.4444444
#>  [49,] 0.8888889 0.4444444
#>  [50,] 1.0000000 0.4444444
#>  [51,] 0.0000000 0.5555556
#>  [52,] 0.1111111 0.5555556
#>  [53,] 0.2222222 0.5555556
#>  [54,] 0.3333333 0.5555556
#>  [55,] 0.4444444 0.5555556
#>  [56,] 0.5555556 0.5555556
#>  [57,] 0.6666667 0.5555556
#>  [58,] 0.7777778 0.5555556
#>  [59,] 0.8888889 0.5555556
#>  [60,] 1.0000000 0.5555556
#>  [61,] 0.0000000 0.6666667
#>  [62,] 0.1111111 0.6666667
#>  [63,] 0.2222222 0.6666667
#>  [64,] 0.3333333 0.6666667
#>  [65,] 0.4444444 0.6666667
#>  [66,] 0.5555556 0.6666667
#>  [67,] 0.6666667 0.6666667
#>  [68,] 0.7777778 0.6666667
#>  [69,] 0.8888889 0.6666667
#>  [70,] 1.0000000 0.6666667
#>  [71,] 0.0000000 0.7777778
#>  [72,] 0.1111111 0.7777778
#>  [73,] 0.2222222 0.7777778
#>  [74,] 0.3333333 0.7777778
#>  [75,] 0.4444444 0.7777778
#>  [76,] 0.5555556 0.7777778
#>  [77,] 0.6666667 0.7777778
#>  [78,] 0.7777778 0.7777778
#>  [79,] 0.8888889 0.7777778
#>  [80,] 1.0000000 0.7777778
#>  [81,] 0.0000000 0.8888889
#>  [82,] 0.1111111 0.8888889
#>  [83,] 0.2222222 0.8888889
#>  [84,] 0.3333333 0.8888889
#>  [85,] 0.4444444 0.8888889
#>  [86,] 0.5555556 0.8888889
#>  [87,] 0.6666667 0.8888889
#>  [88,] 0.7777778 0.8888889
#>  [89,] 0.8888889 0.8888889
#>  [90,] 1.0000000 0.8888889
#>  [91,] 0.0000000 1.0000000
#>  [92,] 0.1111111 1.0000000
#>  [93,] 0.2222222 1.0000000
#>  [94,] 0.3333333 1.0000000
#>  [95,] 0.4444444 1.0000000
#>  [96,] 0.5555556 1.0000000
#>  [97,] 0.6666667 1.0000000
#>  [98,] 0.7777778 1.0000000
#>  [99,] 0.8888889 1.0000000
#> [100,] 1.0000000 1.0000000
#> 
#> $mean
#>              [,1]       [,2]       [,3]
#>   [1,] 0.24102386 0.15180565 0.60717049
#>   [2,] 0.20859621 0.09973560 0.69166820
#>   [3,] 0.18598039 0.11735868 0.69666093
#>   [4,] 0.20206931 0.21491314 0.58301756
#>   [5,] 0.29714849 0.32339828 0.37945323
#>   [6,] 0.32032034 0.31731856 0.36236110
#>   [7,] 0.32585663 0.32740038 0.34674299
#>   [8,] 0.30000904 0.38618075 0.31381022
#>   [9,] 0.30196854 0.40534545 0.29268601
#>  [10,] 0.32096867 0.40027889 0.27875244
#>  [11,] 0.27310249 0.29835982 0.42853769
#>  [12,] 0.25567663 0.27645647 0.46786690
#>  [13,] 0.18072389 0.20816510 0.61111102
#>  [14,] 0.15301700 0.18358825 0.66339475
#>  [15,] 0.23960237 0.23587193 0.52452569
#>  [16,] 0.29228649 0.24348327 0.46423024
#>  [17,] 0.32689753 0.25778421 0.41531826
#>  [18,] 0.37921017 0.33469924 0.28609060
#>  [19,] 0.40202317 0.37807178 0.21990505
#>  [20,] 0.42330937 0.40604560 0.17064503
#>  [21,] 0.22728213 0.30114558 0.47157229
#>  [22,] 0.23042852 0.28952684 0.48004464
#>  [23,] 0.17979601 0.22543054 0.59477346
#>  [24,] 0.05147943 0.06110828 0.88741229
#>  [25,] 0.03130233 0.02933858 0.93935910
#>  [26,] 0.07147240 0.06093318 0.86759442
#>  [27,] 0.23186826 0.21688258 0.55124917
#>  [28,] 0.29750482 0.29895167 0.40354350
#>  [29,] 0.36061535 0.37943284 0.25995181
#>  [30,] 0.38596167 0.41306273 0.20097560
#>  [31,] 0.27642105 0.23343369 0.49014526
#>  [32,] 0.25016949 0.19950645 0.55032406
#>  [33,] 0.17393185 0.14791674 0.67815141
#>  [34,] 0.04537506 0.05060071 0.90402423
#>  [35,] 0.02262952 0.02493968 0.95243080
#>  [36,] 0.03512830 0.03819765 0.92667405
#>  [37,] 0.08343824 0.08252628 0.83403548
#>  [38,] 0.20900636 0.20039796 0.59059567
#>  [39,] 0.31847544 0.34022756 0.34129700
#>  [40,] 0.34253503 0.38628313 0.27118184
#>  [41,] 0.36292706 0.39126306 0.24580988
#>  [42,] 0.28774069 0.30025373 0.41200558
#>  [43,] 0.09409808 0.15163885 0.75426306
#>  [44,] 0.07094863 0.09975431 0.82929706
#>  [45,] 0.09527425 0.07865573 0.82607003
#>  [46,] 0.08716504 0.08207611 0.83075885
#>  [47,] 0.06709260 0.07890016 0.85400724
#>  [48,] 0.11721491 0.11215845 0.77062664
#>  [49,] 0.20997126 0.19931370 0.59071504
#>  [50,] 0.25590605 0.24639285 0.49770110
#>  [51,] 0.39686990 0.41559774 0.18753236
#>  [52,] 0.33647701 0.33576745 0.32775554
#>  [53,] 0.24603378 0.27317575 0.48079047
#>  [54,] 0.20948600 0.23508983 0.55542416
#>  [55,] 0.15954294 0.17043253 0.67002453
#>  [56,] 0.11721905 0.11533325 0.76744770
#>  [57,] 0.10197289 0.09995968 0.79806743
#>  [58,] 0.08007793 0.06656954 0.85335253
#>  [59,] 0.06927978 0.04968673 0.88103349
#>  [60,] 0.09465259 0.07935531 0.82599210
#>  [61,] 0.25211454 0.29874799 0.44913747
#>  [62,] 0.24101270 0.30294439 0.45604291
#>  [63,] 0.26062105 0.35144783 0.38793112
#>  [64,] 0.27727115 0.34936511 0.37336374
#>  [65,] 0.23814701 0.27494351 0.48690947
#>  [66,] 0.19010606 0.22051922 0.58937472
#>  [67,] 0.13921741 0.16475535 0.69602724
#>  [68,] 0.11235418 0.12577664 0.76186918
#>  [69,] 0.11242748 0.10988833 0.77768419
#>  [70,] 0.13458061 0.12557777 0.73984161
#>  [71,] 0.33369840 0.51339806 0.15290354
#>  [72,] 0.33949955 0.53243987 0.12806058
#>  [73,] 0.33268799 0.54224858 0.12506343
#>  [74,] 0.33988239 0.51723162 0.14288598
#>  [75,] 0.35976296 0.38479612 0.25544092
#>  [76,] 0.27976985 0.27458637 0.44564378
#>  [77,] 0.22222775 0.23189849 0.54587376
#>  [78,] 0.13962646 0.17405464 0.68631890
#>  [79,] 0.12446483 0.17534812 0.70018705
#>  [80,] 0.16268385 0.21560881 0.62170734
#>  [81,] 0.47440579 0.46986247 0.05573174
#>  [82,] 0.43901237 0.50248863 0.05849900
#>  [83,] 0.44513446 0.50132351 0.05354203
#>  [84,] 0.46048470 0.47890931 0.06060599
#>  [85,] 0.45789372 0.43680057 0.10530571
#>  [86,] 0.40824782 0.38592758 0.20582461
#>  [87,] 0.33614108 0.34253409 0.32132483
#>  [88,] 0.28454364 0.26560388 0.44985248
#>  [89,] 0.26422550 0.25242026 0.48335424
#>  [90,] 0.25233630 0.24869568 0.49896802
#>  [91,] 0.48226275 0.49101775 0.02671950
#>  [92,] 0.45900637 0.51465108 0.02634255
#>  [93,] 0.48896206 0.47674395 0.03429399
#>  [94,] 0.53299688 0.43308731 0.03391580
#>  [95,] 0.50801250 0.42676996 0.06521753
#>  [96,] 0.37198292 0.46725720 0.16075988
#>  [97,] 0.33703936 0.48653170 0.17642894
#>  [98,] 0.35536832 0.38321521 0.26141646
#>  [99,] 0.33428776 0.27832802 0.38738422
#> [100,] 0.30106038 0.27509502 0.42384460
#> 
#> $variance
#>                [,1]         [,2]         [,3]
#>   [1,] 0.0166721104 0.0117350714 0.0217378797
#>   [2,] 0.0050492929 0.0027462894 0.0065229216
#>   [3,] 0.0041659407 0.0028504310 0.0058151493
#>   [4,] 0.0053364893 0.0055843264 0.0080461760
#>   [5,] 0.0029090333 0.0030477713 0.0032797772
#>   [6,] 0.0014675755 0.0014602433 0.0015574999
#>   [7,] 0.0015274336 0.0015311555 0.0015749809
#>   [8,] 0.0013228827 0.0014932265 0.0013564565
#>   [9,] 0.0025491106 0.0029150233 0.0025036072
#>  [10,] 0.0156712339 0.0172608730 0.0144561872
#>  [11,] 0.0067714063 0.0071406020 0.0083532726
#>  [12,] 0.0025146840 0.0026431521 0.0032898290
#>  [13,] 0.0015836139 0.0017629745 0.0025418459
#>  [14,] 0.0011637039 0.0013458053 0.0020050307
#>  [15,] 0.0013549462 0.0013403944 0.0018547441
#>  [16,] 0.0011020491 0.0009813465 0.0013250930
#>  [17,] 0.0013456547 0.0011701117 0.0014850510
#>  [18,] 0.0011418766 0.0010801084 0.0009906980
#>  [19,] 0.0014749834 0.0014426674 0.0010525297
#>  [20,] 0.0050493993 0.0049884640 0.0029273391
#>  [21,] 0.0021048620 0.0025223223 0.0029865599
#>  [22,] 0.0013768719 0.0015971469 0.0019380101
#>  [23,] 0.0016818471 0.0019913964 0.0027487425
#>  [24,] 0.0003748440 0.0004404388 0.0007669842
#>  [25,] 0.0001561441 0.0001466451 0.0002933310
#>  [26,] 0.0005320484 0.0004587417 0.0009209604
#>  [27,] 0.0014602127 0.0013924854 0.0020281138
#>  [28,] 0.0015163347 0.0015205708 0.0017463321
#>  [29,] 0.0025329485 0.0025866857 0.0021133566
#>  [30,] 0.0037819517 0.0038688690 0.0025625933
#>  [31,] 0.0035162595 0.0031458438 0.0043933434
#>  [32,] 0.0013298125 0.0011321599 0.0017543293
#>  [33,] 0.0011346478 0.0009953262 0.0017236313
#>  [34,] 0.0003530405 0.0003915434 0.0007071575
#>  [35,] 0.0001175590 0.0001292539 0.0002408134
#>  [36,] 0.0002120284 0.0002298211 0.0004250618
#>  [37,] 0.0008790702 0.0008703272 0.0015910962
#>  [38,] 0.0024163311 0.0023420229 0.0035340007
#>  [39,] 0.0014838022 0.0015345541 0.0015368824
#>  [40,] 0.0013652961 0.0014372193 0.0011981992
#>  [41,] 0.0033669945 0.0034684258 0.0026996911
#>  [42,] 0.0023702128 0.0024298358 0.0028017167
#>  [43,] 0.0008233381 0.0012425319 0.0017902330
#>  [44,] 0.0004255613 0.0005797905 0.0009139649
#>  [45,] 0.0003301268 0.0002775496 0.0005502747
#>  [46,] 0.0003062996 0.0002900249 0.0005412436
#>  [47,] 0.0004133018 0.0004798867 0.0008232790
#>  [48,] 0.0009223880 0.0008876531 0.0015756611
#>  [49,] 0.0009241960 0.0008891210 0.0013469924
#>  [50,] 0.0013678448 0.0013338335 0.0017958056
#>  [51,] 0.0017365529 0.0017620325 0.0011053788
#>  [52,] 0.0014483824 0.0014468737 0.0014293847
#>  [53,] 0.0009327854 0.0009984048 0.0012552597
#>  [54,] 0.0005498342 0.0005970510 0.0008198564
#>  [55,] 0.0005422216 0.0005717260 0.0008940382
#>  [56,] 0.0006323968 0.0006235521 0.0010907066
#>  [57,] 0.0009250257 0.0009087961 0.0016278920
#>  [58,] 0.0008454311 0.0007131352 0.0014362082
#>  [59,] 0.0007818554 0.0005725429 0.0012709192
#>  [60,] 0.0027823211 0.0023720701 0.0046666404
#>  [61,] 0.0029271048 0.0032522536 0.0038408541
#>  [62,] 0.0009684750 0.0011180065 0.0013133617
#>  [63,] 0.0008125480 0.0009611213 0.0010012150
#>  [64,] 0.0012891018 0.0014622581 0.0015050635
#>  [65,] 0.0016293197 0.0017902155 0.0022435318
#>  [66,] 0.0010211534 0.0011400364 0.0016051072
#>  [67,] 0.0005267538 0.0006048865 0.0009299971
#>  [68,] 0.0004668928 0.0005147670 0.0008493453
#>  [69,] 0.0008181643 0.0008019740 0.0014175482
#>  [70,] 0.0040510387 0.0038193656 0.0066947422
#>  [71,] 0.0043172915 0.0048508121 0.0025149931
#>  [72,] 0.0017465866 0.0019390359 0.0008697203
#>  [73,] 0.0014979291 0.0016747629 0.0007382987
#>  [74,] 0.0022275541 0.0024791463 0.0012159242
#>  [75,] 0.0035728612 0.0036720505 0.0029501801
#>  [76,] 0.0014052578 0.0013891479 0.0017229020
#>  [77,] 0.0007787709 0.0008025563 0.0011169347
#>  [78,] 0.0005774837 0.0006910698 0.0010349022
#>  [79,] 0.0007322190 0.0009716112 0.0014105393
#>  [80,] 0.0027165950 0.0033727971 0.0046903463
#>  [81,] 0.0032347123 0.0032314275 0.0006827051
#>  [82,] 0.0018794782 0.0019078160 0.0004203165
#>  [83,] 0.0015538493 0.0015727760 0.0003188057
#>  [84,] 0.0018077625 0.0018158877 0.0004142721
#>  [85,] 0.0027682784 0.0027435068 0.0010507205
#>  [86,] 0.0044526125 0.0043679390 0.0030127624
#>  [87,] 0.0056047065 0.0056563012 0.0054772396
#>  [88,] 0.0029122488 0.0027903664 0.0035403461
#>  [89,] 0.0010390048 0.0010085091 0.0013346165
#>  [90,] 0.0012526476 0.0012405865 0.0016598967
#>  [91,] 0.0018774221 0.0018791811 0.0001955398
#>  [92,] 0.0019121416 0.0019234289 0.0001975028
#>  [93,] 0.0029426540 0.0029377197 0.0003900084
#>  [94,] 0.0028680378 0.0028289942 0.0003775352
#>  [95,] 0.0054923798 0.0053759457 0.0013396983
#>  [96,] 0.0095548556 0.0101813007 0.0055181510
#>  [97,] 0.0090172705 0.0100816475 0.0058637794
#>  [98,] 0.0149341635 0.0154087334 0.0125870252
#>  [99,] 0.0120370908 0.0108645387 0.0128364407
#> [100,] 0.0152518259 0.0144541442 0.0177000650
#> 
#> $lower
#>               [,1]        [,2]        [,3]
#>   [1,] 0.045004197 0.012662914 0.305656090
#>   [2,] 0.088383373 0.023017916 0.523330066
#>   [3,] 0.077780408 0.034961714 0.537865450
#>   [4,] 0.080078984 0.088759097 0.403238145
#>   [5,] 0.197288042 0.220326944 0.270811739
#>   [6,] 0.247681243 0.244906498 0.286907615
#>   [7,] 0.251706747 0.253136420 0.271128796
#>   [8,] 0.231250081 0.311921919 0.243983649
#>   [9,] 0.207943122 0.302017596 0.199768424
#>  [10,] 0.107192347 0.162498344 0.081021800
#>  [11,] 0.128734495 0.147903294 0.255665526
#>  [12,] 0.164010392 0.181767151 0.356593148
#>  [13,] 0.109635114 0.132189232 0.510268974
#>  [14,] 0.092422155 0.117398879 0.573031012
#>  [15,] 0.171336230 0.168048807 0.439904032
#>  [16,] 0.229428178 0.184797279 0.393329050
#>  [17,] 0.257132964 0.193694377 0.340879606
#>  [18,] 0.314170821 0.271891589 0.226462367
#>  [19,] 0.328001375 0.305161796 0.159738200
#>  [20,] 0.287778382 0.272056619 0.078783862
#>  [21,] 0.144015250 0.207614540 0.365397235
#>  [22,] 0.161897618 0.214494602 0.394202310
#>  [23,] 0.106821367 0.144328160 0.490214585
#>  [24,] 0.020512512 0.026787229 0.827778057
#>  [25,] 0.011690315 0.010508912 0.901742937
#>  [26,] 0.033202690 0.026080630 0.802797367
#>  [27,] 0.161387707 0.148402350 0.462339208
#>  [28,] 0.224143928 0.225463809 0.323110079
#>  [29,] 0.265174999 0.282550054 0.175216792
#>  [30,] 0.269395980 0.294276666 0.111577661
#>  [31,] 0.168403561 0.133219862 0.361077626
#>  [32,] 0.182241820 0.137811656 0.467678033
#>  [33,] 0.113067618 0.091652443 0.594265061
#>  [34,] 0.016087373 0.019257892 0.846007922
#>  [35,] 0.006538348 0.007784191 0.917752519
#>  [36,] 0.012486136 0.014335338 0.881488711
#>  [37,] 0.035091953 0.034492405 0.748959528
#>  [38,] 0.121383002 0.114521213 0.471929581
#>  [39,] 0.245482697 0.265666414 0.266663473
#>  [40,] 0.272045216 0.313399010 0.206107299
#>  [41,] 0.253448675 0.279279125 0.151587218
#>  [42,] 0.197348766 0.208384805 0.310488842
#>  [43,] 0.045795374 0.089334126 0.666910437
#>  [44,] 0.036076932 0.057725257 0.766178750
#>  [45,] 0.062723890 0.049196733 0.777803015
#>  [46,] 0.056004643 0.051880398 0.782826328
#>  [47,] 0.032993479 0.041541652 0.793529294
#>  [48,] 0.064562269 0.060731116 0.688468599
#>  [49,] 0.153593358 0.144194371 0.517909252
#>  [50,] 0.186931367 0.178459013 0.414803831
#>  [51,] 0.316759230 0.334628635 0.126897193
#>  [52,] 0.264032543 0.263370744 0.255911382
#>  [53,] 0.188705124 0.213511739 0.411608448
#>  [54,] 0.165420034 0.188925731 0.498994874
#>  [55,] 0.116615263 0.126192418 0.610174245
#>  [56,] 0.072591740 0.071075494 0.699744336
#>  [57,] 0.050505676 0.049060729 0.713559431
#>  [58,] 0.032922191 0.024402873 0.771805074
#>  [59,] 0.025171823 0.014016420 0.802928293
#>  [60,] 0.019284520 0.012916387 0.673599981
#>  [61,] 0.154021949 0.193483831 0.329613788
#>  [62,] 0.182738874 0.239493194 0.385537998
#>  [63,] 0.206746468 0.291940177 0.326873706
#>  [64,] 0.209768132 0.276385248 0.298996922
#>  [65,] 0.163772747 0.196104960 0.394479555
#>  [66,] 0.131569734 0.158038712 0.509838389
#>  [67,] 0.097364853 0.119458154 0.634689728
#>  [68,] 0.073597536 0.084773974 0.702508661
#>  [69,] 0.062752304 0.060813945 0.699769640
#>  [70,] 0.037066817 0.032337121 0.565010144
#>  [71,] 0.211822103 0.376982443 0.068593594
#>  [72,] 0.260138601 0.445797327 0.076066065
#>  [73,] 0.259117103 0.461613175 0.076858907
#>  [74,] 0.250656104 0.419526443 0.081698286
#>  [75,] 0.247251683 0.269910537 0.156821660
#>  [76,] 0.209379838 0.204689253 0.365137741
#>  [77,] 0.170003795 0.178760448 0.480041122
#>  [78,] 0.095953556 0.125627487 0.621643996
#>  [79,] 0.076466750 0.118602502 0.624150413
#>  [80,] 0.074504344 0.113599372 0.483442087
#>  [81,] 0.363890032 0.359516860 0.016381549
#>  [82,] 0.355074176 0.416984884 0.025132057
#>  [83,] 0.368639488 0.423683914 0.024204890
#>  [84,] 0.377832197 0.395813163 0.027168552
#>  [85,] 0.355925098 0.335747682 0.050716766
#>  [86,] 0.281296307 0.261030531 0.109485213
#>  [87,] 0.198397335 0.203796888 0.186001264
#>  [88,] 0.185147401 0.168981313 0.334986082
#>  [89,] 0.203554158 0.192805882 0.412003591
#>  [90,] 0.186253790 0.182995269 0.419242553
#>  [91,] 0.397731624 0.406321017 0.006589257
#>  [92,] 0.374047746 0.428620643 0.006261275
#>  [93,] 0.383164614 0.371312146 0.006889397
#>  [94,] 0.427569827 0.330588394 0.006895785
#>  [95,] 0.363138356 0.286919931 0.013472750
#>  [96,] 0.192800034 0.274090259 0.045296687
#>  [97,] 0.165987723 0.292713480 0.054728581
#>  [98,] 0.139386483 0.159706003 0.077795627
#>  [99,] 0.140532473 0.101559619 0.180721975
#> [100,] 0.093900515 0.078284485 0.179470646
#> 
#> $upper
#>              [,1]       [,2]       [,3]
#>   [1,] 0.53363531 0.41746989 0.86818093
#>   [2,] 0.36349181 0.22341961 0.83711264
#>   [3,] 0.32767576 0.24026845 0.83438802
#>   [4,] 0.36249400 0.37787407 0.75202361
#>   [5,] 0.40786274 0.43592110 0.49454844
#>   [6,] 0.39757958 0.39442789 0.44135420
#>   [7,] 0.40462647 0.40624336 0.42642332
#>   [8,] 0.37357109 0.46317382 0.38810868
#>   [9,] 0.40517748 0.51306522 0.39521683
#>  [10,] 0.58702614 0.66727000 0.54090729
#>  [11,] 0.44776658 0.47576052 0.61097815
#>  [12,] 0.35973013 0.38248246 0.58077185
#>  [13,] 0.26486766 0.29608066 0.70740035
#>  [14,] 0.22549786 0.26062418 0.74814898
#>  [15,] 0.31525709 0.31118923 0.60845089
#>  [16,] 0.35935967 0.30737320 0.53585749
#>  [17,] 0.40069785 0.32751937 0.49173158
#>  [18,] 0.44648048 0.40055970 0.34966879
#>  [19,] 0.47833711 0.45383410 0.28661965
#>  [20,] 0.56498141 0.54755741 0.28874883
#>  [21,] 0.32307028 0.40381452 0.57905456
#>  [22,] 0.30694693 0.37079831 0.56647881
#>  [23,] 0.26673957 0.31852215 0.69518831
#>  [24,] 0.09546784 0.10819553 0.93572700
#>  [25,] 0.06000184 0.05728550 0.96840225
#>  [26,] 0.12277925 0.10912279 0.92116337
#>  [27,] 0.31074143 0.29422298 0.63855346
#>  [28,] 0.37647230 0.37800609 0.48664847
#>  [29,] 0.46192286 0.48139109 0.35478400
#>  [30,] 0.50953693 0.53719319 0.30870753
#>  [31,] 0.39958730 0.35169415 0.61988152
#>  [32,] 0.32485280 0.26932284 0.63160836
#>  [33,] 0.24461615 0.21477839 0.75666486
#>  [34,] 0.08864882 0.09579017 0.94951585
#>  [35,] 0.04821954 0.05156578 0.97801402
#>  [36,] 0.06872265 0.07295276 0.96173182
#>  [37,] 0.14999635 0.14880957 0.90442889
#>  [38,] 0.31295532 0.30307956 0.70416803
#>  [39,] 0.39620228 0.41895585 0.42006977
#>  [40,] 0.41666408 0.46179566 0.34154322
#>  [41,] 0.48008213 0.50933761 0.35425117
#>  [42,] 0.38753977 0.40097620 0.51742473
#>  [43,] 0.15731982 0.22678609 0.83222567
#>  [44,] 0.11632899 0.15160630 0.88431752
#>  [45,] 0.13370834 0.11423619 0.86959050
#>  [46,] 0.12435624 0.11837573 0.87385199
#>  [47,] 0.11203192 0.12681528 0.90558469
#>  [48,] 0.18286704 0.17675353 0.84356890
#>  [49,] 0.27250202 0.26081136 0.66159484
#>  [50,] 0.33156987 0.32127604 0.58066141
#>  [51,] 0.47983671 0.49890442 0.25681002
#>  [52,] 0.41296732 0.41222753 0.40386120
#>  [53,] 0.30822400 0.33718234 0.55034039
#>  [54,] 0.25721740 0.28459671 0.61115383
#>  [55,] 0.20770345 0.21973868 0.72725925
#>  [56,] 0.17074345 0.16853137 0.82892309
#>  [57,] 0.16875301 0.16624658 0.87106553
#>  [58,] 0.14555486 0.12760133 0.91940363
#>  [59,] 0.13320650 0.10595436 0.94150160
#>  [60,] 0.22044761 0.19777736 0.93723933
#>  [61,] 0.36500362 0.41603276 0.57170181
#>  [62,] 0.30450754 0.37036923 0.52743446
#>  [63,] 0.31833549 0.41333898 0.45078677
#>  [64,] 0.35023691 0.42604056 0.45083756
#>  [65,] 0.32150596 0.36150678 0.57978901
#>  [66,] 0.25647431 0.29006539 0.66664966
#>  [67,] 0.18709767 0.21565556 0.75408581
#>  [68,] 0.15800597 0.17343789 0.81656443
#>  [69,] 0.17420108 0.17113883 0.84690753
#>  [70,] 0.28132669 0.26917680 0.88201256
#>  [71,] 0.46805161 0.64880772 0.26312315
#>  [72,] 0.42363460 0.61811724 0.19108920
#>  [73,] 0.41056532 0.62179630 0.18289579
#>  [74,] 0.43519285 0.61428181 0.21760518
#>  [75,] 0.48064804 0.50656170 0.36864712
#>  [76,] 0.35601886 0.35048022 0.52759660
#>  [77,] 0.27921315 0.28963249 0.61091959
#>  [78,] 0.18988589 0.22844200 0.74758400
#>  [79,] 0.18206578 0.24040500 0.77109324
#>  [80,] 0.27675073 0.33949482 0.75058555
#>  [81,] 0.58619679 0.58170969 0.11686711
#>  [82,] 0.52472808 0.58791984 0.10462682
#>  [83,] 0.52294584 0.57893134 0.09351018
#>  [84,] 0.54423495 0.56259137 0.10615741
#>  [85,] 0.56166240 0.54055519 0.17667195
#>  [86,] 0.54173043 0.51894359 0.32304356
#>  [87,] 0.48987807 0.49664183 0.47408405
#>  [88,] 0.39578254 0.37510706 0.56747790
#>  [89,] 0.32969565 0.31707340 0.55504384
#>  [90,] 0.32468861 0.32075791 0.57871963
#>  [91,] 0.56730318 0.57597241 0.06015344
#>  [92,] 0.54517067 0.60025061 0.06004851
#>  [93,] 0.59525813 0.58322630 0.08219027
#>  [94,] 0.63696591 0.53854276 0.08100251
#>  [95,] 0.65220398 0.57285834 0.15312904
#>  [96,] 0.57184708 0.66571871 0.33038990
#>  [97,] 0.53403496 0.68249768 0.34919150
#>  [98,] 0.60946512 0.63752724 0.50757885
#>  [99,] 0.56387133 0.50290508 0.61842914
#> [100,] 0.56674079 0.53793315 0.69072955
#> 
#> $CI_level
#> [1] 0.95
#>