Predict Method for BKP or DKP Models

Generates posterior predictive summaries from a fitted BKP or DKP model at new input locations.

Usage

# S3 method for class 'BKP'
predict(object, Xnew, CI_level = 0.95, threshold = 0.5, ...)

# S3 method for class 'DKP'
predict(object, Xnew, CI_level = 0.95, ...)

Arguments

object

An object of class "BKP" or "DKP", typically returned by fit.BKP or fit.DKP.

Xnew

A numeric matrix (or vector) of new input locations where predictions are desired.

CI_level

Credible level for prediction intervals (default is 0.95, corresponding to 95% CI).

threshold

Classification threshold for binary prediction based on posterior mean (used only for BKP; default is 0.5).

...

Additional arguments passed to generic predict methods (currently not used; included for S3 method consistency).

Value

A list with the following components:

Xnew

The new input locations.

mean

BKP: Posterior mean of the success probability at each location. DKP: A matrix of posterior mean class probabilities (rows = inputs, columns = classes).

variance

BKP: Posterior variance of the success probability. DKP: A matrix of posterior variances for each class.

lower

BKP: Lower bound of the prediction interval (e.g., 2.5th percentile for 95% CI). DKP: A matrix of lower bounds for each class (e.g., 2.5th percentile).

upper

BKP: Upper bound of the prediction interval (e.g., 97.5th percentile for 95% CI). DKP: A matrix of upper bounds for each class (e.g., 97.5th percentile).

class

BKP: Predicted binary label (0 or 1), based on posterior mean and threshold, if m = 1. DKP: Predicted class label (i.e., the class with the highest posterior mean probability).

See also

Examples

# ============================================================== #
# ========================= BKP Examples ======================= #
# ============================================================== #

#-------------------------- 1D Example ---------------------------
set.seed(123)

# Define true success probability function
true_pi_fun <- function(x) {
  (1 + exp(-x^2) * cos(10 * (1 - exp(-x)) / (1 + exp(-x)))) / 2
}

n <- 30
Xbounds <- matrix(c(-2,2), nrow=1)
X <- tgp::lhs(n = n, rect = Xbounds)
true_pi <- true_pi_fun(X)
m <- sample(100, n, replace = TRUE)
y <- rbinom(n, size = m, prob = true_pi)

# Fit BKP model
model1 <- fit.BKP(X, y, m, Xbounds=Xbounds)

# Prediction
Xnew = matrix(seq(-2, 2, length = 10), ncol=1) #new data points
predict(model1, Xnew)
#> $Xnew
#>             [,1]
#>  [1,] -2.0000000
#>  [2,] -1.5555556
#>  [3,] -1.1111111
#>  [4,] -0.6666667
#>  [5,] -0.2222222
#>  [6,]  0.2222222
#>  [7,]  0.6666667
#>  [8,]  1.1111111
#>  [9,]  1.5555556
#> [10,]  2.0000000
#> 
#> $mean
#>  [1] 0.4876541 0.5318748 0.5175070 0.1956923 0.8602008 0.6057005 0.2259782
#>  [8] 0.5370140 0.5700729 0.5110550
#> 
#> $variance
#>  [1] 0.008132938 0.001834905 0.002690709 0.001167453 0.001886007 0.002055462
#>  [7] 0.001327923 0.002292589 0.002096727 0.006504398
#> 
#> $lower
#>  [1] 0.3130318 0.4476040 0.4157174 0.1332364 0.7650464 0.5152804 0.1587223
#>  [8] 0.4427054 0.4793517 0.3533662
#> 
#> $upper
#>  [1] 0.6638517 0.6152488 0.6185737 0.2667556 0.9337123 0.6926431 0.3011716
#>  [8] 0.6300170 0.6585024 0.6676235
#> 
#> $CI_level
#> [1] 0.95
#> 

#-------------------------- 2D Example ---------------------------
set.seed(123)

# Define 2D latent function and probability transformation
true_pi_fun <- function(X) {
  if(is.null(nrow(X))) X <- matrix(X, nrow=1)
  m <- 8.6928
  s <- 2.4269
  x1 <- 4*X[,1]- 2
  x2 <- 4*X[,2]- 2
  a <- 1 + (x1 + x2 + 1)^2 *
    (19- 14*x1 + 3*x1^2- 14*x2 + 6*x1*x2 + 3*x2^2)
  b <- 30 + (2*x1- 3*x2)^2 *
    (18- 32*x1 + 12*x1^2 + 48*x2- 36*x1*x2 + 27*x2^2)
  f <- log(a*b)
  f <- (f- m)/s
  return(pnorm(f))  # Transform to probability
}

n <- 100
Xbounds <- matrix(c(0, 0, 1, 1), nrow = 2)
X <- tgp::lhs(n = n, rect = Xbounds)
true_pi <- true_pi_fun(X)
m <- sample(100, n, replace = TRUE)
y <- rbinom(n, size = m, prob = true_pi)

# Fit BKP model
model2 <- fit.BKP(X, y, m, Xbounds=Xbounds)

# Prediction
x1 <- seq(Xbounds[1,1], Xbounds[1,2], length.out = 10)
x2 <- seq(Xbounds[2,1], Xbounds[2,2], length.out = 10)
Xnew <- expand.grid(x1 = x1, x2 = x2)
predict(model2, Xnew)
#> $Xnew
#>               x1        x2
#>   [1,] 0.0000000 0.0000000
#>   [2,] 0.1111111 0.0000000
#>   [3,] 0.2222222 0.0000000
#>   [4,] 0.3333333 0.0000000
#>   [5,] 0.4444444 0.0000000
#>   [6,] 0.5555556 0.0000000
#>   [7,] 0.6666667 0.0000000
#>   [8,] 0.7777778 0.0000000
#>   [9,] 0.8888889 0.0000000
#>  [10,] 1.0000000 0.0000000
#>  [11,] 0.0000000 0.1111111
#>  [12,] 0.1111111 0.1111111
#>  [13,] 0.2222222 0.1111111
#>  [14,] 0.3333333 0.1111111
#>  [15,] 0.4444444 0.1111111
#>  [16,] 0.5555556 0.1111111
#>  [17,] 0.6666667 0.1111111
#>  [18,] 0.7777778 0.1111111
#>  [19,] 0.8888889 0.1111111
#>  [20,] 1.0000000 0.1111111
#>  [21,] 0.0000000 0.2222222
#>  [22,] 0.1111111 0.2222222
#>  [23,] 0.2222222 0.2222222
#>  [24,] 0.3333333 0.2222222
#>  [25,] 0.4444444 0.2222222
#>  [26,] 0.5555556 0.2222222
#>  [27,] 0.6666667 0.2222222
#>  [28,] 0.7777778 0.2222222
#>  [29,] 0.8888889 0.2222222
#>  [30,] 1.0000000 0.2222222
#>  [31,] 0.0000000 0.3333333
#>  [32,] 0.1111111 0.3333333
#>  [33,] 0.2222222 0.3333333
#>  [34,] 0.3333333 0.3333333
#>  [35,] 0.4444444 0.3333333
#>  [36,] 0.5555556 0.3333333
#>  [37,] 0.6666667 0.3333333
#>  [38,] 0.7777778 0.3333333
#>  [39,] 0.8888889 0.3333333
#>  [40,] 1.0000000 0.3333333
#>  [41,] 0.0000000 0.4444444
#>  [42,] 0.1111111 0.4444444
#>  [43,] 0.2222222 0.4444444
#>  [44,] 0.3333333 0.4444444
#>  [45,] 0.4444444 0.4444444
#>  [46,] 0.5555556 0.4444444
#>  [47,] 0.6666667 0.4444444
#>  [48,] 0.7777778 0.4444444
#>  [49,] 0.8888889 0.4444444
#>  [50,] 1.0000000 0.4444444
#>  [51,] 0.0000000 0.5555556
#>  [52,] 0.1111111 0.5555556
#>  [53,] 0.2222222 0.5555556
#>  [54,] 0.3333333 0.5555556
#>  [55,] 0.4444444 0.5555556
#>  [56,] 0.5555556 0.5555556
#>  [57,] 0.6666667 0.5555556
#>  [58,] 0.7777778 0.5555556
#>  [59,] 0.8888889 0.5555556
#>  [60,] 1.0000000 0.5555556
#>  [61,] 0.0000000 0.6666667
#>  [62,] 0.1111111 0.6666667
#>  [63,] 0.2222222 0.6666667
#>  [64,] 0.3333333 0.6666667
#>  [65,] 0.4444444 0.6666667
#>  [66,] 0.5555556 0.6666667
#>  [67,] 0.6666667 0.6666667
#>  [68,] 0.7777778 0.6666667
#>  [69,] 0.8888889 0.6666667
#>  [70,] 1.0000000 0.6666667
#>  [71,] 0.0000000 0.7777778
#>  [72,] 0.1111111 0.7777778
#>  [73,] 0.2222222 0.7777778
#>  [74,] 0.3333333 0.7777778
#>  [75,] 0.4444444 0.7777778
#>  [76,] 0.5555556 0.7777778
#>  [77,] 0.6666667 0.7777778
#>  [78,] 0.7777778 0.7777778
#>  [79,] 0.8888889 0.7777778
#>  [80,] 1.0000000 0.7777778
#>  [81,] 0.0000000 0.8888889
#>  [82,] 0.1111111 0.8888889
#>  [83,] 0.2222222 0.8888889
#>  [84,] 0.3333333 0.8888889
#>  [85,] 0.4444444 0.8888889
#>  [86,] 0.5555556 0.8888889
#>  [87,] 0.6666667 0.8888889
#>  [88,] 0.7777778 0.8888889
#>  [89,] 0.8888889 0.8888889
#>  [90,] 1.0000000 0.8888889
#>  [91,] 0.0000000 1.0000000
#>  [92,] 0.1111111 1.0000000
#>  [93,] 0.2222222 1.0000000
#>  [94,] 0.3333333 1.0000000
#>  [95,] 0.4444444 1.0000000
#>  [96,] 0.5555556 1.0000000
#>  [97,] 0.6666667 1.0000000
#>  [98,] 0.7777778 1.0000000
#>  [99,] 0.8888889 1.0000000
#> [100,] 1.0000000 1.0000000
#> 
#> $mean
#>   [1] 0.40856269 0.37382617 0.34487626 0.37952496 0.60268953 0.61168110
#>   [7] 0.60379956 0.62802429 0.66109943 0.65567569 0.54269012 0.51383695
#>  [13] 0.36343491 0.31972347 0.49084529 0.55898705 0.58445974 0.71066278
#>  [19] 0.77310238 0.82870175 0.59231102 0.55598137 0.42309320 0.09848559
#>  [25] 0.05107664 0.11089917 0.44302895 0.54972800 0.72583877 0.83724446
#>  [31] 0.56100858 0.51328066 0.39241450 0.09202271 0.03508809 0.05556137
#>  [37] 0.14063382 0.34547061 0.60414014 0.67639035 0.81451893 0.65070363
#>  [43] 0.19266967 0.13040729 0.15287792 0.15362112 0.13666652 0.19515769
#>  [49] 0.37751283 0.48173994 0.78582219 0.65751633 0.48179292 0.41828036
#>  [55] 0.30896813 0.21665439 0.18458735 0.11826157 0.08893491 0.14417018
#>  [61] 0.53424792 0.53740971 0.60580451 0.59523734 0.48147372 0.39242325
#>  [67] 0.28454083 0.23722495 0.21411630 0.22921402 0.81801698 0.85790372
#>  [73] 0.87624925 0.88369041 0.76579808 0.53895837 0.47832614 0.33501640
#>  [79] 0.31878295 0.39496156 0.95721165 0.94122149 0.95500241 0.95051703
#>  [85] 0.88400223 0.78596072 0.67398670 0.49677274 0.48966037 0.48761845
#>  [91] 0.98497065 0.98677687 0.97577669 0.96850692 0.93591606 0.78657512
#>  [97] 0.75054243 0.63561906 0.52518822 0.50825512
#> 
#> $variance
#>   [1] 0.0312611948 0.0078141682 0.0072498334 0.0107700929 0.0043034924
#>   [6] 0.0017580221 0.0020239873 0.0015725978 0.0032655071 0.0279870950
#>  [11] 0.0114859122 0.0037108574 0.0029290774 0.0022103459 0.0022777943
#>  [16] 0.0014428122 0.0018722697 0.0011111192 0.0012030139 0.0039557781
#>  [21] 0.0032894835 0.0020914065 0.0036081950 0.0008046743 0.0002697550
#>  [26] 0.0010011134 0.0023594702 0.0020072781 0.0027324090 0.0024801210
#>  [31] 0.0055956427 0.0018900182 0.0021937067 0.0008418265 0.0002017021
#>  [36] 0.0003811630 0.0017403305 0.0045181091 0.0018704125 0.0014274643
#>  [41] 0.0024874585 0.0031485613 0.0017795669 0.0008953149 0.0005514560
#>  [46] 0.0005615286 0.0009203975 0.0018193603 0.0014458084 0.0020133597
#>  [51] 0.0013125428 0.0017416461 0.0015408000 0.0008763942 0.0010164899
#>  [56] 0.0012304168 0.0018908966 0.0014583181 0.0010845975 0.0053978765
#>  [61] 0.0050764012 0.0014701925 0.0011267045 0.0019022194 0.0027476709
#>  [66] 0.0019014142 0.0009963313 0.0009662251 0.0015874024 0.0092560866
#>  [71] 0.0036890531 0.0010694298 0.0008312625 0.0011936142 0.0038317390
#>  [76] 0.0020136266 0.0012474057 0.0012307637 0.0016833362 0.0065437395
#>  [81] 0.0006326648 0.0004955107 0.0003061622 0.0003937090 0.0013302868
#>  [86] 0.0038461397 0.0071673045 0.0049035277 0.0014104197 0.0018118337
#>  [91] 0.0001184951 0.0001121307 0.0003508356 0.0003884406 0.0016373787
#>  [96] 0.0087589588 0.0083134032 0.0211184902 0.0170571955 0.0243417828
#> 
#> $lower
#>   [1] 0.10168696 0.21031254 0.18917418 0.18950977 0.47110230 0.52806649
#>   [7] 0.51411793 0.54878129 0.54500896 0.30026487 0.33110281 0.39443874
#>  [13] 0.26100998 0.23133558 0.39764802 0.48397662 0.49855696 0.64326456
#>  [19] 0.70172059 0.68934978 0.47783847 0.46560962 0.30794747 0.05023493
#>  [25] 0.02390634 0.05688132 0.34906902 0.46130225 0.61786704 0.72874101
#>  [31] 0.41247783 0.42801983 0.30270615 0.04349057 0.01286199 0.02382232
#>  [37] 0.06946257 0.22038323 0.51797314 0.60027355 0.70753632 0.53710277
#>  [43] 0.11712037 0.07761199 0.10976654 0.11012726 0.08288363 0.11872334
#>  [49] 0.30453289 0.39423025 0.71077487 0.57355574 0.40517478 0.36086384
#>  [55] 0.24829215 0.15201068 0.10742425 0.05434207 0.03566210 0.03406736
#>  [61] 0.39389537 0.46193311 0.53913664 0.50847296 0.37937372 0.30875598
#>  [67] 0.22477115 0.17908547 0.14146348 0.07254340 0.68508030 0.78814805
#>  [73] 0.81453749 0.80803199 0.63443379 0.45056032 0.40938345 0.26809039
#>  [79] 0.24121780 0.24308428 0.89608550 0.89065919 0.91503716 0.90504512
#>  [85] 0.80366264 0.65285519 0.49846350 0.36025185 0.41625314 0.40449317
#>  [91] 0.95741155 0.95946343 0.92783660 0.92000511 0.83674311 0.57738029
#>  [97] 0.55267031 0.33302033 0.27028892 0.20924117
#> 
#> $upper
#>   [1] 0.76618226 0.55385228 0.52005879 0.59141140 0.72715292 0.69213985
#>   [7] 0.69012453 0.70397931 0.76816106 0.92904970 0.74642372 0.63244415
#>  [13] 0.47249645 0.41512396 0.58436172 0.63268145 0.66787574 0.77372497
#>  [19] 0.83734284 0.93257568 0.70195123 0.64454021 0.54261327 0.16058519
#>  [25] 0.08770173 0.17997316 0.53907361 0.63661284 0.82190368 0.92215806
#>  [31] 0.70416134 0.59815779 0.48590921 0.15619651 0.06774872 0.09952145
#>  [37] 0.23161464 0.48249646 0.68719615 0.74812008 0.90160665 0.75628221
#>  [43] 0.28167309 0.19431824 0.20160794 0.20280167 0.20124770 0.28510575
#>  [49] 0.45336615 0.56981228 0.85236447 0.73682473 0.55883954 0.47681624
#>  [55] 0.37310637 0.28913259 0.27692417 0.20258156 0.16316348 0.31497256
#>  [61] 0.67189602 0.61204288 0.67057180 0.67912936 0.58435522 0.47936790
#>  [67] 0.34832494 0.30070732 0.29707471 0.44217267 0.92047851 0.91568685
#>  [73] 0.92697797 0.94234739 0.87521868 0.62615007 0.54768149 0.40541512
#>  [79] 0.40171191 0.55804910 0.99187625 0.97681853 0.98272565 0.98173821
#>  [85] 0.94531978 0.89378550 0.82729500 0.63353846 0.56329003 0.57108651
#>  [91] 0.99836120 0.99900479 0.99795901 0.99482069 0.99015129 0.93654391
#>  [97] 0.90450751 0.88699974 0.77308485 0.80390905
#> 
#> $CI_level
#> [1] 0.95
#> 

# ============================================================== #
# ========================= DKP Examples ======================= #
# ============================================================== #

#-------------------------- 1D Example ---------------------------
set.seed(123)

# Define true class probability function (3-class)
true_pi_fun <- function(X) {
  p <- (1 + exp(-X^2) * cos(10 * (1 - exp(-X)) / (1 + exp(-X)))) / 2
  return(matrix(c(p/2, p/2, 1 - p), nrow = length(p)))
}

n <- 30
Xbounds <- matrix(c(-2, 2), nrow = 1)
X <- tgp::lhs(n = n, rect = Xbounds)
true_pi <- true_pi_fun(X)
m <- sample(100, n, replace = TRUE)

# Generate multinomial responses
Y <- t(sapply(1:n, function(i) rmultinom(1, size = m[i], prob = true_pi[i, ])))

# Fit DKP model
model1 <- fit.DKP(X, Y, Xbounds = Xbounds)

# Prediction
Xnew = matrix(seq(-2, 2, length = 10), ncol=1) #new data points
predict(model1, Xnew)
#> $Xnew
#>             [,1]
#>  [1,] -2.0000000
#>  [2,] -1.5555556
#>  [3,] -1.1111111
#>  [4,] -0.6666667
#>  [5,] -0.2222222
#>  [6,]  0.2222222
#>  [7,]  0.6666667
#>  [8,]  1.1111111
#>  [9,]  1.5555556
#> [10,]  2.0000000
#> 
#> $mean
#>            [,1]       [,2]      [,3]
#>  [1,] 0.1919444 0.27657759 0.5314780
#>  [2,] 0.3065415 0.29965790 0.3938006
#>  [3,] 0.3193745 0.27097374 0.4096518
#>  [4,] 0.1112803 0.06597305 0.8227467
#>  [5,] 0.3907534 0.43166186 0.1775848
#>  [6,] 0.2753590 0.32248139 0.4021596
#>  [7,] 0.1341606 0.09007242 0.7757670
#>  [8,] 0.3060871 0.23434495 0.4595679
#>  [9,] 0.2569524 0.27407645 0.4689711
#> [10,] 0.2974804 0.32737373 0.3751459
#> 
#> $variance
#>               [,1]         [,2]        [,3]
#>  [1,] 0.0045401038 0.0058567672 0.007288939
#>  [2,] 0.0014509468 0.0014324442 0.001629422
#>  [3,] 0.0021779426 0.0019792851 0.002423043
#>  [4,] 0.0007050916 0.0004393275 0.001039736
#>  [5,] 0.0034093227 0.0035133605 0.002091554
#>  [6,] 0.0016557605 0.0018130144 0.001995074
#>  [7,] 0.0008172041 0.0005765897 0.001223768
#>  [8,] 0.0018732590 0.0015824740 0.002190477
#>  [9,] 0.0015368291 0.0016014702 0.002004567
#> [10,] 0.0049852170 0.0052527283 0.005591729
#> 
#> $lower
#>             [,1]       [,2]       [,3]
#>  [1,] 0.07906969 0.14070981 0.36348464
#>  [2,] 0.23454336 0.22822864 0.31619251
#>  [3,] 0.23161605 0.18843642 0.31512798
#>  [4,] 0.06478408 0.03112775 0.75537129
#>  [5,] 0.27970993 0.31774730 0.09739606
#>  [6,] 0.19936718 0.24202548 0.31633381
#>  [7,] 0.08327339 0.04879376 0.70368081
#>  [8,] 0.22473409 0.16109480 0.36869548
#>  [9,] 0.18405193 0.19930662 0.38184954
#> [10,] 0.16943484 0.19424026 0.23523976
#> 
#> $upper
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3398352 0.4379332 0.6958718
#>  [2,] 0.3835768 0.3763034 0.4741746
#>  [3,] 0.4140591 0.3622901 0.5076413
#>  [4,] 0.1683209 0.1125517 0.8813517
#>  [5,] 0.5078130 0.5493403 0.2754764
#>  [6,] 0.3584707 0.4085653 0.4910881
#>  [7,] 0.1948475 0.1423104 0.8404550
#>  [8,] 0.3939771 0.3165454 0.5518040
#>  [9,] 0.3373225 0.3557905 0.5570471
#> [10,] 0.4442648 0.4764876 0.5266162
#> 
#> $CI_level
#> [1] 0.95
#> 

#-------------------------- 2D Example ---------------------------
set.seed(123)

# Define latent function and transform to 3-class probabilities
true_pi_fun <- function(X) {
  if (is.null(nrow(X))) X <- matrix(X, nrow = 1)
  m <- 8.6928; s <- 2.4269
  x1 <- 4 * X[,1] - 2
  x2 <- 4 * X[,2] - 2
  a <- 1 + (x1 + x2 + 1)^2 *
    (19 - 14*x1 + 3*x1^2 - 14*x2 + 6*x1*x2 + 3*x2^2)
  b <- 30 + (2*x1 - 3*x2)^2 *
    (18 - 32*x1 + 12*x1^2 + 48*x2 - 36*x1*x2 + 27*x2^2)
  f <- (log(a * b) - m) / s
  p <- pnorm(f)
  return(matrix(c(p/2, p/2, 1 - p), nrow = length(p)))
}

n <- 100
Xbounds <- matrix(c(0, 0, 1, 1), nrow = 2)
X <- tgp::lhs(n = n, rect = Xbounds)
true_pi <- true_pi_fun(X)
m <- sample(100, n, replace = TRUE)

# Generate multinomial responses
Y <- t(sapply(1:n, function(i) rmultinom(1, size = m[i], prob = true_pi[i, ])))

# Fit DKP model
model2 <- fit.DKP(X, Y, Xbounds = Xbounds)
print(model2)
#> --------------------------------------------------
#>            Dirichlet Kernel Process (DKP) Model        
#> --------------------------------------------------
#> Number of observations (n):  100
#> Input dimensionality (d):    2
#> Output dimensionality (q):   3
#> Kernel type:                 gaussian
#> Loss function used:          brier
#> Optimized kernel parameters: 0.1245, 0.0706
#> Minimum achieved loss:       0.01060
#> Kernel parameters were obtained by optimization.
#> 
#> Prior specification:
#>   Noninformative prior: Dirichlet(1,...,1).
#> --------------------------------------------------

# Prediction
x1 <- seq(Xbounds[1,1], Xbounds[1,2], length.out = 10)
x2 <- seq(Xbounds[2,1], Xbounds[2,2], length.out = 10)
Xnew <- expand.grid(x1 = x1, x2 = x2)
predict(model2, Xnew)
#> $Xnew
#>               x1        x2
#>   [1,] 0.0000000 0.0000000
#>   [2,] 0.1111111 0.0000000
#>   [3,] 0.2222222 0.0000000
#>   [4,] 0.3333333 0.0000000
#>   [5,] 0.4444444 0.0000000
#>   [6,] 0.5555556 0.0000000
#>   [7,] 0.6666667 0.0000000
#>   [8,] 0.7777778 0.0000000
#>   [9,] 0.8888889 0.0000000
#>  [10,] 1.0000000 0.0000000
#>  [11,] 0.0000000 0.1111111
#>  [12,] 0.1111111 0.1111111
#>  [13,] 0.2222222 0.1111111
#>  [14,] 0.3333333 0.1111111
#>  [15,] 0.4444444 0.1111111
#>  [16,] 0.5555556 0.1111111
#>  [17,] 0.6666667 0.1111111
#>  [18,] 0.7777778 0.1111111
#>  [19,] 0.8888889 0.1111111
#>  [20,] 1.0000000 0.1111111
#>  [21,] 0.0000000 0.2222222
#>  [22,] 0.1111111 0.2222222
#>  [23,] 0.2222222 0.2222222
#>  [24,] 0.3333333 0.2222222
#>  [25,] 0.4444444 0.2222222
#>  [26,] 0.5555556 0.2222222
#>  [27,] 0.6666667 0.2222222
#>  [28,] 0.7777778 0.2222222
#>  [29,] 0.8888889 0.2222222
#>  [30,] 1.0000000 0.2222222
#>  [31,] 0.0000000 0.3333333
#>  [32,] 0.1111111 0.3333333
#>  [33,] 0.2222222 0.3333333
#>  [34,] 0.3333333 0.3333333
#>  [35,] 0.4444444 0.3333333
#>  [36,] 0.5555556 0.3333333
#>  [37,] 0.6666667 0.3333333
#>  [38,] 0.7777778 0.3333333
#>  [39,] 0.8888889 0.3333333
#>  [40,] 1.0000000 0.3333333
#>  [41,] 0.0000000 0.4444444
#>  [42,] 0.1111111 0.4444444
#>  [43,] 0.2222222 0.4444444
#>  [44,] 0.3333333 0.4444444
#>  [45,] 0.4444444 0.4444444
#>  [46,] 0.5555556 0.4444444
#>  [47,] 0.6666667 0.4444444
#>  [48,] 0.7777778 0.4444444
#>  [49,] 0.8888889 0.4444444
#>  [50,] 1.0000000 0.4444444
#>  [51,] 0.0000000 0.5555556
#>  [52,] 0.1111111 0.5555556
#>  [53,] 0.2222222 0.5555556
#>  [54,] 0.3333333 0.5555556
#>  [55,] 0.4444444 0.5555556
#>  [56,] 0.5555556 0.5555556
#>  [57,] 0.6666667 0.5555556
#>  [58,] 0.7777778 0.5555556
#>  [59,] 0.8888889 0.5555556
#>  [60,] 1.0000000 0.5555556
#>  [61,] 0.0000000 0.6666667
#>  [62,] 0.1111111 0.6666667
#>  [63,] 0.2222222 0.6666667
#>  [64,] 0.3333333 0.6666667
#>  [65,] 0.4444444 0.6666667
#>  [66,] 0.5555556 0.6666667
#>  [67,] 0.6666667 0.6666667
#>  [68,] 0.7777778 0.6666667
#>  [69,] 0.8888889 0.6666667
#>  [70,] 1.0000000 0.6666667
#>  [71,] 0.0000000 0.7777778
#>  [72,] 0.1111111 0.7777778
#>  [73,] 0.2222222 0.7777778
#>  [74,] 0.3333333 0.7777778
#>  [75,] 0.4444444 0.7777778
#>  [76,] 0.5555556 0.7777778
#>  [77,] 0.6666667 0.7777778
#>  [78,] 0.7777778 0.7777778
#>  [79,] 0.8888889 0.7777778
#>  [80,] 1.0000000 0.7777778
#>  [81,] 0.0000000 0.8888889
#>  [82,] 0.1111111 0.8888889
#>  [83,] 0.2222222 0.8888889
#>  [84,] 0.3333333 0.8888889
#>  [85,] 0.4444444 0.8888889
#>  [86,] 0.5555556 0.8888889
#>  [87,] 0.6666667 0.8888889
#>  [88,] 0.7777778 0.8888889
#>  [89,] 0.8888889 0.8888889
#>  [90,] 1.0000000 0.8888889
#>  [91,] 0.0000000 1.0000000
#>  [92,] 0.1111111 1.0000000
#>  [93,] 0.2222222 1.0000000
#>  [94,] 0.3333333 1.0000000
#>  [95,] 0.4444444 1.0000000
#>  [96,] 0.5555556 1.0000000
#>  [97,] 0.6666667 1.0000000
#>  [98,] 0.7777778 1.0000000
#>  [99,] 0.8888889 1.0000000
#> [100,] 1.0000000 1.0000000
#> 
#> $mean
#>              [,1]       [,2]       [,3]
#>   [1,] 0.24102382 0.15180559 0.60717059
#>   [2,] 0.20859620 0.09973559 0.69166821
#>   [3,] 0.18598040 0.11735868 0.69666092
#>   [4,] 0.20206936 0.21491317 0.58301747
#>   [5,] 0.29714850 0.32339828 0.37945323
#>   [6,] 0.32032035 0.31731856 0.36236109
#>   [7,] 0.32585662 0.32740041 0.34674298
#>   [8,] 0.30000904 0.38618073 0.31381023
#>   [9,] 0.30196853 0.40534544 0.29268603
#>  [10,] 0.32096864 0.40027890 0.27875246
#>  [11,] 0.27310249 0.29835981 0.42853770
#>  [12,] 0.25567660 0.27645645 0.46786694
#>  [13,] 0.18072389 0.20816510 0.61111101
#>  [14,] 0.15301704 0.18358828 0.66339468
#>  [15,] 0.23960238 0.23587193 0.52452569
#>  [16,] 0.29228649 0.24348326 0.46423025
#>  [17,] 0.32689755 0.25778423 0.41531823
#>  [18,] 0.37921017 0.33469923 0.28609060
#>  [19,] 0.40202317 0.37807176 0.21990507
#>  [20,] 0.42330936 0.40604558 0.17064506
#>  [21,] 0.22728213 0.30114558 0.47157229
#>  [22,] 0.23042851 0.28952684 0.48004465
#>  [23,] 0.17979597 0.22543050 0.59477352
#>  [24,] 0.05147944 0.06110828 0.88741228
#>  [25,] 0.03130233 0.02933858 0.93935909
#>  [26,] 0.07147242 0.06093321 0.86759437
#>  [27,] 0.23186824 0.21688256 0.55124920
#>  [28,] 0.29750481 0.29895167 0.40354352
#>  [29,] 0.36061532 0.37943281 0.25995187
#>  [30,] 0.38596166 0.41306272 0.20097562
#>  [31,] 0.27642104 0.23343366 0.49014531
#>  [32,] 0.25016947 0.19950643 0.55032409
#>  [33,] 0.17393185 0.14791673 0.67815142
#>  [34,] 0.04537509 0.05060072 0.90402420
#>  [35,] 0.02262952 0.02493968 0.95243079
#>  [36,] 0.03512830 0.03819764 0.92667406
#>  [37,] 0.08343823 0.08252626 0.83403550
#>  [38,] 0.20900639 0.20039801 0.59059560
#>  [39,] 0.31847544 0.34022758 0.34129698
#>  [40,] 0.34253503 0.38628313 0.27118184
#>  [41,] 0.36292704 0.39126304 0.24580993
#>  [42,] 0.28774063 0.30025371 0.41200566
#>  [43,] 0.09409810 0.15163887 0.75426303
#>  [44,] 0.07094864 0.09975430 0.82929706
#>  [45,] 0.09527424 0.07865573 0.82607003
#>  [46,] 0.08716504 0.08207611 0.83075885
#>  [47,] 0.06709262 0.07890016 0.85400722
#>  [48,] 0.11721493 0.11215848 0.77062659
#>  [49,] 0.20997126 0.19931370 0.59071504
#>  [50,] 0.25590603 0.24639283 0.49770115
#>  [51,] 0.39686988 0.41559773 0.18753239
#>  [52,] 0.33647701 0.33576746 0.32775552
#>  [53,] 0.24603379 0.27317575 0.48079046
#>  [54,] 0.20948600 0.23508983 0.55542418
#>  [55,] 0.15954295 0.17043255 0.67002450
#>  [56,] 0.11721906 0.11533327 0.76744768
#>  [57,] 0.10197289 0.09995967 0.79806744
#>  [58,] 0.08007793 0.06656954 0.85335254
#>  [59,] 0.06927978 0.04968673 0.88103349
#>  [60,] 0.09465255 0.07935527 0.82599217
#>  [61,] 0.25211451 0.29874796 0.44913752
#>  [62,] 0.24101271 0.30294440 0.45604289
#>  [63,] 0.26062104 0.35144781 0.38793115
#>  [64,] 0.27727113 0.34936511 0.37336376
#>  [65,] 0.23814701 0.27494353 0.48690946
#>  [66,] 0.19010604 0.22051920 0.58937476
#>  [67,] 0.13921741 0.16475535 0.69602725
#>  [68,] 0.11235419 0.12577665 0.76186916
#>  [69,] 0.11242748 0.10988834 0.77768418
#>  [70,] 0.13458059 0.12557775 0.73984166
#>  [71,] 0.33369841 0.51339808 0.15290351
#>  [72,] 0.33949954 0.53243988 0.12806057
#>  [73,] 0.33268799 0.54224859 0.12506343
#>  [74,] 0.33988239 0.51723162 0.14288598
#>  [75,] 0.35976290 0.38479611 0.25544099
#>  [76,] 0.27976983 0.27458637 0.44564380
#>  [77,] 0.22222774 0.23189849 0.54587378
#>  [78,] 0.13962648 0.17405466 0.68631885
#>  [79,] 0.12446483 0.17534812 0.70018705
#>  [80,] 0.16268380 0.21560878 0.62170742
#>  [81,] 0.47440577 0.46986248 0.05573175
#>  [82,] 0.43901238 0.50248862 0.05849900
#>  [83,] 0.44513447 0.50132350 0.05354203
#>  [84,] 0.46048470 0.47890931 0.06060599
#>  [85,] 0.45789371 0.43680059 0.10530570
#>  [86,] 0.40824784 0.38592760 0.20582456
#>  [87,] 0.33614110 0.34253408 0.32132482
#>  [88,] 0.28454365 0.26560388 0.44985247
#>  [89,] 0.26422551 0.25242025 0.48335424
#>  [90,] 0.25233630 0.24869569 0.49896801
#>  [91,] 0.48226274 0.49101776 0.02671950
#>  [92,] 0.45900638 0.51465107 0.02634255
#>  [93,] 0.48896206 0.47674395 0.03429399
#>  [94,] 0.53299688 0.43308732 0.03391580
#>  [95,] 0.50801252 0.42676997 0.06521751
#>  [96,] 0.37198297 0.46725719 0.16075983
#>  [97,] 0.33703937 0.48653169 0.17642894
#>  [98,] 0.35536831 0.38321526 0.26141644
#>  [99,] 0.33428778 0.27832804 0.38738418
#> [100,] 0.30106041 0.27509503 0.42384456
#> 
#> $variance
#>                [,1]         [,2]         [,3]
#>   [1,] 0.0166720977 0.0117350594 0.0217378632
#>   [2,] 0.0050492923 0.0027462891 0.0065229209
#>   [3,] 0.0041659405 0.0028504307 0.0058151488
#>   [4,] 0.0053364884 0.0055843249 0.0080461734
#>   [5,] 0.0029090322 0.0030477702 0.0032797761
#>   [6,] 0.0014675755 0.0014602433 0.0015574999
#>   [7,] 0.0015274330 0.0015311550 0.0015749802
#>   [8,] 0.0013228826 0.0014932263 0.0013564564
#>   [9,] 0.0025491096 0.0029150222 0.0025036064
#>  [10,] 0.0156712143 0.0172608526 0.0144561707
#>  [11,] 0.0067714036 0.0071405990 0.0083532692
#>  [12,] 0.0025146838 0.0026431519 0.0032898290
#>  [13,] 0.0015836135 0.0017629740 0.0025418452
#>  [14,] 0.0011637039 0.0013458052 0.0020050304
#>  [15,] 0.0013549457 0.0013403939 0.0018547434
#>  [16,] 0.0011020491 0.0009813465 0.0013250929
#>  [17,] 0.0013456542 0.0011701113 0.0014850504
#>  [18,] 0.0011418765 0.0010801083 0.0009906980
#>  [19,] 0.0014749832 0.0014426672 0.0010525296
#>  [20,] 0.0050493955 0.0049884603 0.0029273374
#>  [21,] 0.0021048614 0.0025223217 0.0029865592
#>  [22,] 0.0013768719 0.0015971469 0.0019380101
#>  [23,] 0.0016818460 0.0019913952 0.0027487410
#>  [24,] 0.0003748439 0.0004404387 0.0007669841
#>  [25,] 0.0001561441 0.0001466451 0.0002933311
#>  [26,] 0.0005320483 0.0004587417 0.0009209603
#>  [27,] 0.0014602122 0.0013924850 0.0020281132
#>  [28,] 0.0015163345 0.0015205707 0.0017463320
#>  [29,] 0.0025329478 0.0025866849 0.0021133564
#>  [30,] 0.0037819515 0.0038688688 0.0025625934
#>  [31,] 0.0035162577 0.0031458419 0.0043933413
#>  [32,] 0.0013298124 0.0011321597 0.0017543292
#>  [33,] 0.0011346474 0.0009953259 0.0017236307
#>  [34,] 0.0003530405 0.0003915434 0.0007071574
#>  [35,] 0.0001175590 0.0001292539 0.0002408134
#>  [36,] 0.0002120284 0.0002298210 0.0004250617
#>  [37,] 0.0008790698 0.0008703268 0.0015910956
#>  [38,] 0.0024163301 0.0023420221 0.0035339991
#>  [39,] 0.0014838020 0.0015345539 0.0015368822
#>  [40,] 0.0013652961 0.0014372193 0.0011981992
#>  [41,] 0.0033669944 0.0034684257 0.0026996914
#>  [42,] 0.0023702121 0.0024298353 0.0028017164
#>  [43,] 0.0008233381 0.0012425316 0.0017902326
#>  [44,] 0.0004255612 0.0005797903 0.0009139645
#>  [45,] 0.0003301267 0.0002775495 0.0005502746
#>  [46,] 0.0003062996 0.0002900248 0.0005412435
#>  [47,] 0.0004133018 0.0004798866 0.0008232788
#>  [48,] 0.0009223877 0.0008876528 0.0015756605
#>  [49,] 0.0009241960 0.0008891210 0.0013469924
#>  [50,] 0.0013678445 0.0013338332 0.0017958054
#>  [51,] 0.0017365527 0.0017620323 0.0011053788
#>  [52,] 0.0014483820 0.0014468733 0.0014293843
#>  [53,] 0.0009327851 0.0009984044 0.0012552593
#>  [54,] 0.0005498341 0.0005970510 0.0008198563
#>  [55,] 0.0005422215 0.0005717259 0.0008940380
#>  [56,] 0.0006323967 0.0006235521 0.0010907065
#>  [57,] 0.0009250255 0.0009087958 0.0016278915
#>  [58,] 0.0008454309 0.0007131350 0.0014362078
#>  [59,] 0.0007818554 0.0005725429 0.0012709191
#>  [60,] 0.0027823187 0.0023720678 0.0046666364
#>  [61,] 0.0029271030 0.0032522516 0.0038408520
#>  [62,] 0.0009684749 0.0011180064 0.0013133615
#>  [63,] 0.0008125479 0.0009611212 0.0010012150
#>  [64,] 0.0012891015 0.0014622577 0.0015050632
#>  [65,] 0.0016293191 0.0017902149 0.0022435309
#>  [66,] 0.0010211530 0.0011400360 0.0016051067
#>  [67,] 0.0005267538 0.0006048865 0.0009299970
#>  [68,] 0.0004668928 0.0005147669 0.0008493452
#>  [69,] 0.0008181642 0.0008019738 0.0014175479
#>  [70,] 0.0040510345 0.0038193616 0.0066947353
#>  [71,] 0.0043172897 0.0048508100 0.0025149916
#>  [72,] 0.0017465865 0.0019390357 0.0008697201
#>  [73,] 0.0014979290 0.0016747627 0.0007382986
#>  [74,] 0.0022275535 0.0024791457 0.0012159239
#>  [75,] 0.0035728585 0.0036720480 0.0029501786
#>  [76,] 0.0014052574 0.0013891476 0.0017229016
#>  [77,] 0.0007787707 0.0008025562 0.0011169346
#>  [78,] 0.0005774837 0.0006910697 0.0010349021
#>  [79,] 0.0007322189 0.0009716111 0.0014105392
#>  [80,] 0.0027165932 0.0033727953 0.0046903439
#>  [81,] 0.0032347120 0.0032314272 0.0006827050
#>  [82,] 0.0018794779 0.0019078158 0.0004203165
#>  [83,] 0.0015538490 0.0015727758 0.0003188057
#>  [84,] 0.0018077621 0.0018158873 0.0004142720
#>  [85,] 0.0027682776 0.0027435061 0.0010507202
#>  [86,] 0.0044526114 0.0043679380 0.0030127611
#>  [87,] 0.0056047056 0.0056563000 0.0054772383
#>  [88,] 0.0029122470 0.0027903646 0.0035403438
#>  [89,] 0.0010390047 0.0010085090 0.0013346164
#>  [90,] 0.0012526474 0.0012405862 0.0016598964
#>  [91,] 0.0018774220 0.0018791809 0.0001955398
#>  [92,] 0.0019121412 0.0019234285 0.0001975027
#>  [93,] 0.0029426528 0.0029377185 0.0003900081
#>  [94,] 0.0028680377 0.0028289941 0.0003775352
#>  [95,] 0.0054923777 0.0053759437 0.0013396975
#>  [96,] 0.0095548518 0.0101812961 0.0055181472
#>  [97,] 0.0090172702 0.0100816471 0.0058637793
#>  [98,] 0.0149341576 0.0154087283 0.0125870197
#>  [99,] 0.0120370954 0.0108645432 0.0128364450
#> [100,] 0.0152518310 0.0144541486 0.0177000696
#> 
#> $lower
#>               [,1]        [,2]        [,3]
#>   [1,] 0.045004223 0.012662915 0.305656285
#>   [2,] 0.088383372 0.023017916 0.523330087
#>   [3,] 0.077780419 0.034961716 0.537865451
#>   [4,] 0.080079041 0.088759140 0.403238088
#>   [5,] 0.197288063 0.220326956 0.270811758
#>   [6,] 0.247681246 0.244906505 0.286907604
#>   [7,] 0.251706747 0.253136458 0.271128798
#>   [8,] 0.231250087 0.311921911 0.243983658
#>   [9,] 0.207943127 0.302017607 0.199768456
#>  [10,] 0.107192425 0.162498478 0.081021902
#>  [11,] 0.128734521 0.147903315 0.255665566
#>  [12,] 0.164010372 0.181767133 0.356593193
#>  [13,] 0.109635127 0.132189248 0.510268978
#>  [14,] 0.092422189 0.117398908 0.573030959
#>  [15,] 0.171336247 0.168048813 0.439904047
#>  [16,] 0.229428177 0.184797276 0.393329060
#>  [17,] 0.257132990 0.193694405 0.340879591
#>  [18,] 0.314170829 0.271891581 0.226462375
#>  [19,] 0.328001381 0.305161784 0.159738220
#>  [20,] 0.287778417 0.272056647 0.078783914
#>  [21,] 0.144015259 0.207614549 0.365397251
#>  [22,] 0.161897608 0.214494601 0.394202324
#>  [23,] 0.106821357 0.144328149 0.490214681
#>  [24,] 0.020512521 0.026787239 0.827778050
#>  [25,] 0.011690318 0.010508915 0.901742924
#>  [26,] 0.033202711 0.026080652 0.802797325
#>  [27,] 0.161387696 0.148402349 0.462339255
#>  [28,] 0.224143921 0.225463806 0.323110099
#>  [29,] 0.265174984 0.282550037 0.175216853
#>  [30,] 0.269395979 0.294276653 0.111577677
#>  [31,] 0.168403575 0.133219855 0.361077703
#>  [32,] 0.182241808 0.137811645 0.467678066
#>  [33,] 0.113067627 0.091652445 0.594265083
#>  [34,] 0.016087391 0.019257904 0.846007891
#>  [35,] 0.006538348 0.007784193 0.917752515
#>  [36,] 0.012486135 0.014335338 0.881488720
#>  [37,] 0.035091950 0.034492402 0.748959570
#>  [38,] 0.121383045 0.114521271 0.471929537
#>  [39,] 0.245482706 0.265666429 0.266663461
#>  [40,] 0.272045217 0.313399007 0.206107299
#>  [41,] 0.253448656 0.279279104 0.151587255
#>  [42,] 0.197348721 0.208384793 0.310488925
#>  [43,] 0.045795394 0.089334143 0.666910414
#>  [44,] 0.036076941 0.057725254 0.766178767
#>  [45,] 0.062723886 0.049196737 0.777803024
#>  [46,] 0.056004645 0.051880397 0.782826338
#>  [47,] 0.032993495 0.041541662 0.793529283
#>  [48,] 0.064562299 0.060731149 0.688468566
#>  [49,] 0.153593357 0.144194370 0.517909257
#>  [50,] 0.186931352 0.178458996 0.414803878
#>  [51,] 0.316759225 0.334628623 0.126897219
#>  [52,] 0.264032555 0.263370769 0.255911374
#>  [53,] 0.188705140 0.213511746 0.411608456
#>  [54,] 0.165420031 0.188925725 0.498994893
#>  [55,] 0.116615280 0.126192438 0.610174220
#>  [56,] 0.072591753 0.071075509 0.699744314
#>  [57,] 0.050505680 0.049060728 0.713559456
#>  [58,] 0.032922193 0.024402875 0.771805087
#>  [59,] 0.025171821 0.014016419 0.802928297
#>  [60,] 0.019284514 0.012916380 0.673600120
#>  [61,] 0.154021953 0.193483838 0.329613866
#>  [62,] 0.182738882 0.239493212 0.385537984
#>  [63,] 0.206746466 0.291940159 0.326873732
#>  [64,] 0.209768128 0.276385250 0.298996949
#>  [65,] 0.163772763 0.196104985 0.394479557
#>  [66,] 0.131569727 0.158038708 0.509838438
#>  [67,] 0.097364853 0.119458153 0.634689738
#>  [68,] 0.073597546 0.084773985 0.702508648
#>  [69,] 0.062752309 0.060813953 0.699769643
#>  [70,] 0.037066836 0.032337138 0.565010280
#>  [71,] 0.211822132 0.376982496 0.068593589
#>  [72,] 0.260138600 0.445797344 0.076066059
#>  [73,] 0.259117108 0.461613181 0.076858906
#>  [74,] 0.250656114 0.419526456 0.081698293
#>  [75,] 0.247251664 0.269910570 0.156821745
#>  [76,] 0.209379836 0.204689256 0.365137771
#>  [77,] 0.170003790 0.178760445 0.480041144
#>  [78,] 0.095953582 0.125627512 0.621643950
#>  [79,] 0.076466749 0.118602503 0.624150423
#>  [80,] 0.074504326 0.113599365 0.483442201
#>  [81,] 0.363890024 0.359516875 0.016381553
#>  [82,] 0.355074197 0.416984879 0.025132055
#>  [83,] 0.368639497 0.423683917 0.024204892
#>  [84,] 0.377832201 0.395813172 0.027168556
#>  [85,] 0.355925108 0.335747708 0.050716766
#>  [86,] 0.281296345 0.261030569 0.109485188
#>  [87,] 0.198397367 0.203796893 0.186001266
#>  [88,] 0.185147434 0.168981345 0.334986110
#>  [89,] 0.203554163 0.192805884 0.412003591
#>  [90,] 0.186253802 0.182995276 0.419242553
#>  [91,] 0.397731619 0.406321031 0.006589257
#>  [92,] 0.374047770 0.428620640 0.006261276
#>  [93,] 0.383164639 0.371312173 0.006889396
#>  [94,] 0.427569825 0.330588397 0.006895785
#>  [95,] 0.363138396 0.286919960 0.013472748
#>  [96,] 0.192800113 0.274090298 0.045296679
#>  [97,] 0.165987732 0.292713474 0.054728585
#>  [98,] 0.139386506 0.159706071 0.077795637
#>  [99,] 0.140532451 0.101559609 0.180721914
#> [100,] 0.093900509 0.078284470 0.179470586
#> 
#> $upper
#>              [,1]       [,2]       [,3]
#>   [1,] 0.53363516 0.41746970 0.86818092
#>   [2,] 0.36349179 0.22341960 0.83711264
#>   [3,] 0.32767576 0.24026844 0.83438801
#>   [4,] 0.36249404 0.37787407 0.75202350
#>   [5,] 0.40786272 0.43592107 0.49454842
#>   [6,] 0.39757959 0.39442790 0.44135419
#>   [7,] 0.40462644 0.40624338 0.42642329
#>   [8,] 0.37357109 0.46317380 0.38810869
#>   [9,] 0.40517745 0.51306520 0.39521683
#>  [10,] 0.58702593 0.66726985 0.54090715
#>  [11,] 0.44776654 0.47576047 0.61097812
#>  [12,] 0.35973010 0.38248244 0.58077189
#>  [13,] 0.26486766 0.29608065 0.70740032
#>  [14,] 0.22549789 0.26062420 0.74814891
#>  [15,] 0.31525708 0.31118921 0.60845087
#>  [16,] 0.35935967 0.30737319 0.53585749
#>  [17,] 0.40069785 0.32751937 0.49173153
#>  [18,] 0.44648048 0.40055969 0.34966879
#>  [19,] 0.47833710 0.45383408 0.28661966
#>  [20,] 0.56498134 0.54755733 0.28874882
#>  [21,] 0.32307027 0.40381451 0.57905454
#>  [22,] 0.30694691 0.37079831 0.56647882
#>  [23,] 0.26673951 0.31852209 0.69518835
#>  [24,] 0.09546784 0.10819554 0.93572698
#>  [25,] 0.06000184 0.05728551 0.96840225
#>  [26,] 0.12277927 0.10912281 0.92116333
#>  [27,] 0.31074139 0.29422295 0.63855348
#>  [28,] 0.37647228 0.37800608 0.48664848
#>  [29,] 0.46192282 0.48139104 0.35478405
#>  [30,] 0.50953692 0.53719317 0.30870755
#>  [31,] 0.39958726 0.35169408 0.61988154
#>  [32,] 0.32485278 0.26932282 0.63160838
#>  [33,] 0.24461614 0.21477837 0.75666485
#>  [34,] 0.08864884 0.09579018 0.94951582
#>  [35,] 0.04821954 0.05156579 0.97801402
#>  [36,] 0.06872265 0.07295275 0.96173182
#>  [37,] 0.14999633 0.14880955 0.90442890
#>  [38,] 0.31295533 0.30307958 0.70416793
#>  [39,] 0.39620228 0.41895586 0.42006975
#>  [40,] 0.41666408 0.46179566 0.34154322
#>  [41,] 0.48008211 0.50933759 0.35425122
#>  [42,] 0.38753970 0.40097616 0.51742481
#>  [43,] 0.15731983 0.22678609 0.83222563
#>  [44,] 0.11632898 0.15160628 0.88431752
#>  [45,] 0.13370832 0.11423619 0.86959050
#>  [46,] 0.12435624 0.11837572 0.87385199
#>  [47,] 0.11203193 0.12681528 0.90558466
#>  [48,] 0.18286705 0.17675355 0.84356884
#>  [49,] 0.27250201 0.26081135 0.66159484
#>  [50,] 0.33156985 0.32127601 0.58066145
#>  [51,] 0.47983670 0.49890440 0.25681005
#>  [52,] 0.41296731 0.41222754 0.40386117
#>  [53,] 0.30822400 0.33718232 0.55034038
#>  [54,] 0.25721739 0.28459670 0.61115384
#>  [55,] 0.20770345 0.21973869 0.72725922
#>  [56,] 0.17074346 0.16853138 0.82892306
#>  [57,] 0.16875300 0.16624656 0.87106553
#>  [58,] 0.14555485 0.12760132 0.91940363
#>  [59,] 0.13320649 0.10595436 0.94150160
#>  [60,] 0.22044752 0.19777726 0.93723936
#>  [61,] 0.36500355 0.41603270 0.57170182
#>  [62,] 0.30450754 0.37036924 0.52743444
#>  [63,] 0.31833548 0.41333896 0.45078680
#>  [64,] 0.35023689 0.42604054 0.45083757
#>  [65,] 0.32150594 0.36150678 0.57978898
#>  [66,] 0.25647428 0.29006536 0.66664968
#>  [67,] 0.18709767 0.21565555 0.75408581
#>  [68,] 0.15800598 0.17343790 0.81656441
#>  [69,] 0.17420107 0.17113883 0.84690752
#>  [70,] 0.28132658 0.26917670 0.88201254
#>  [71,] 0.46805159 0.64880771 0.26312309
#>  [72,] 0.42363459 0.61811725 0.19108919
#>  [73,] 0.41056531 0.62179630 0.18289578
#>  [74,] 0.43519283 0.61428180 0.21760517
#>  [75,] 0.48064794 0.50656165 0.36864715
#>  [76,] 0.35601884 0.35048021 0.52759661
#>  [77,] 0.27921314 0.28963247 0.61091961
#>  [78,] 0.18988591 0.22844202 0.74758395
#>  [79,] 0.18206578 0.24040500 0.77109324
#>  [80,] 0.27675064 0.33949475 0.75058560
#>  [81,] 0.58619677 0.58170970 0.11686711
#>  [82,] 0.52472809 0.58791983 0.10462681
#>  [83,] 0.52294583 0.57893133 0.09351018
#>  [84,] 0.54423494 0.56259136 0.10615741
#>  [85,] 0.56166238 0.54055519 0.17667193
#>  [86,] 0.54173044 0.51894360 0.32304349
#>  [87,] 0.48987808 0.49664180 0.47408402
#>  [88,] 0.39578250 0.37510703 0.56747786
#>  [89,] 0.32969565 0.31707339 0.55504384
#>  [90,] 0.32468861 0.32075790 0.57871961
#>  [91,] 0.56730317 0.57597241 0.06015344
#>  [92,] 0.54517067 0.60025059 0.06004850
#>  [93,] 0.59525811 0.58322629 0.08219024
#>  [94,] 0.63696591 0.53854276 0.08100251
#>  [95,] 0.65220397 0.57285831 0.15312900
#>  [96,] 0.57184709 0.66571866 0.33038979
#>  [97,] 0.53403496 0.68249766 0.34919150
#>  [98,] 0.60946505 0.63752723 0.50757876
#>  [99,] 0.56387138 0.50290515 0.61842915
#> [100,] 0.56674086 0.53793321 0.69072955
#> 
#> $CI_level
#> [1] 0.95
#>